Database.Design.Ampersand

Documentation

data A_Context

data P_Context

Constructors

PCtx

Fields

ctx_nm :: String

The name of this context

ctx_pos :: [Origin]

The origin of the context. A context can be a merge of a file including other files c.q. a list of Origin.

ctx_lang :: Lang

The default language specified on the top-level context

ctx_markup :: Maybe PandocFormat

The default markup format for free text in this context

ctx_thms :: [String]

Names of patterns/processes to be printed in the functional specification. (For partial documents.)

ctx_pats :: [P_Pattern]

The patterns defined in this context

ctx_rs :: [P_Rule TermPrim]

All user defined rules in this context, but outside patterns and outside processes

ctx_ds :: [P_Declaration]

The relations defined in this context, outside the scope of patterns

ctx_cs :: [ConceptDef]

The concept definitions defined in this context, outside the scope of patterns

ctx_ks :: [P_IdentDef]

The identity definitions defined in this context, outside the scope of patterns

ctx_rrules :: [P_RoleRule]

The MAINTAIN definitions defined in this context, outside the scope of patterns

ctx_rrels :: [P_RoleRelation]

The assignment of roles to Relations. (EDITS statements)

ctx_vs :: [P_ViewDef]

The view definitions defined in this context, outside the scope of patterns

ctx_gs :: [P_Gen]

The gen definitions defined in this context, outside the scope of patterns

ctx_ifcs :: [P_Interface]

The interfaces defined in this context

ctx_ps :: [PPurpose]

The purposes defined in this context, outside the scope of patterns and processes

ctx_pops :: [P_Population]

The populations defined in this context

ctx_sql :: [P_ObjectDef]

user defined sqlplugs, taken from the Ampersand script

ctx_php :: [P_ObjectDef]

user defined phpplugs, taken from the Ampersand script

ctx_metas :: [Meta]

generic meta information (name/value pairs) that can be used for experimenting without having to modify the adl syntax

data P_Population

Constructors

P_RelPopu

Fields

p_orig :: Origin

p_rnme :: String

p_popps :: Pairs

P_TRelPop

Fields

p_orig :: Origin

p_rnme :: String

p_type :: P_Sign

p_popps :: Pairs

P_CptPopu

Fields

p_orig :: Origin

p_cnme :: String

p_popas :: [String]

data PairView a

Constructors

PairView

Fields

ppv_segs :: [PairViewSegment a]

data PairViewSegment a

Constructors

PairViewText

Fields

pvsOrg :: Origin

pvsStr :: String

PairViewExp

Fields

pvsOrg :: Origin

pvsSoT :: SrcOrTgt

pvsExp :: a

data SrcOrTgt

Constructors

Src

Tgt

data P_Rule a

Constructors

P_Ru

Fields

rr_fps :: Origin

Position in the Ampersand file

rr_nm :: String

Name of this rule

rr_exp :: Term a

The rule expression

rr_mean :: [PMeaning]

User-specified meanings, possibly more than one, for multiple languages.

rr_msg :: [PMessage]

User-specified violation messages, possibly more than one, for multiple languages.

rr_viol :: Maybe (PairView (Term a))

Custom presentation for violations, currently only in a single language

data Term a

Constructors

Prim a

PEqu Origin (Term a) (Term a)
equivalence =

PImp Origin (Term a) (Term a)
implication |-

PIsc Origin (Term a) (Term a)
intersection /

PUni Origin (Term a) (Term a)
union /

PDif Origin (Term a) (Term a)
difference -

PLrs Origin (Term a) (Term a)
left residual /

PRrs Origin (Term a) (Term a)
right residual

PDia Origin (Term a) (Term a)
diamond <>

PCps Origin (Term a) (Term a)
composition ;

PRad Origin (Term a) (Term a)
relative addition !

PPrd Origin (Term a) (Term a)
cartesian product *

PKl0 Origin (Term a)
Rfx.Trn closure * (Kleene star)

PKl1 Origin (Term a)
Transitive closure + (Kleene plus)

PFlp Origin (Term a)
conversion (flip, wok) ~

PCpl Origin (Term a)
Complement

PBrk Origin (Term a)
bracketed expression ( ... )

data TermPrim

Constructors

PI Origin
identity element without a type At parse time, there may be zero or one element in the list of concepts. Reason: when making eqClasses, the least element of that class is used as a witness of that class to know whether an eqClass represents a concept, we only look at its witness By making Pid the first in the data decleration, it becomes the least element for "deriving Ord".

Pid Origin P_Concept
identity element restricted to a type

Patm Origin String (Maybe P_Concept)
an atom, possibly with a type

PVee Origin
the complete relation, of which the type is yet to be derived by the type checker.

Pfull Origin P_Concept P_Concept
the complete relation, restricted to a type. At parse time, there may be zero, one or two elements in the list of concepts.

PNamedR P_NamedRel

data P_NamedRel

Constructors

PNamedRel Origin String (Maybe P_Sign)

data P_Sign

Constructors

P_Sign

Fields

pSrc :: P_Concept

pTgt :: P_Concept

data P_Concept

Constructors

PCpt
The name of this Concept

Fields

p_cptnm :: String

P_Singleton

data P_Declaration

Constructors

P_Sgn

Fields

dec_nm :: String

the name of the declaration

dec_sign :: P_Sign

the type. Parser must guarantee it is not empty.

dec_prps :: Props

the user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx)

dec_pragma :: [String]

Three strings, which form the pragma. E.g. if pragma consists of the three strings: "Person ", " is married to person ", and " in Vegas." ^ then a tuple (Peter,Jane) in the list of links means that Person Peter is married to person Jane in Vegas.

dec_Mean :: [PMeaning]

the optional meaning of a declaration, possibly more than one for different languages.

dec_popu :: Pairs

the list of tuples, of which the relation consists.

dec_fpos :: Origin

the position in the Ampersand source file where this declaration is declared. Not all decalartions come from the ampersand souce file.

dec_plug :: Bool

if true, this relation may not be stored in or retrieved from the standard database (it should be gotten from a Plug of some sort instead)

data P_Pattern

Constructors

P_Pat

Fields

pt_pos :: Origin

the starting position in the file in which this pattern was declared.

pt_nm :: String

Name of this pattern

pt_rls :: [P_Rule TermPrim]

The user defined rules in this pattern

pt_gns :: [P_Gen]

The generalizations defined in this pattern

pt_dcs :: [P_Declaration]

The relations that are declared in this pattern

pt_RRuls :: [P_RoleRule]

The assignment of roles to rules.

pt_RRels :: [P_RoleRelation]

The assignment of roles to Relations.

pt_cds :: [ConceptDef]

The concept definitions defined in this pattern

pt_ids :: [P_IdentDef]

The identity definitions defined in this pattern

pt_vds :: [P_ViewDef]

The view definitions defined in this pattern

pt_xps :: [PPurpose]

The purposes of elements defined in this pattern

pt_pop :: [P_Population]

The populations that are local to this pattern

pt_end :: Origin

the end position in the file in which this pattern was declared.

data P_Gen

Constructors

P_Cy

Fields

gen_fp :: Origin

Position in the Ampersand file

gen_spc :: P_Concept

Left hand side concept expression

gen_rhs :: [P_Concept]

Right hand side concept expression

PGen

Fields

gen_fp :: Origin

Position in the Ampersand file

gen_spc :: P_Concept

Left hand side concept expression

gen_gen :: P_Concept

generic concept

data P_Markup

Constructors

P_Markup

Fields

mLang :: Maybe Lang

mFormat :: Maybe PandocFormat

mString :: String

data PRef2Obj

Constructors

PRef2ConceptDef String

PRef2Declaration P_NamedRel

PRef2Rule String

PRef2IdentityDef String

PRef2ViewDef String

PRef2Pattern String

PRef2Interface String

PRef2Context String

PRef2Fspc String

data PPurpose

Constructors

PRef2

Fields

pexPos :: Origin

pexObj :: PRef2Obj

pexMarkup :: P_Markup

pexRefIDs :: [String]

newtype PMeaning

Constructors

PMeaning P_Markup

data Meta

Constructors

Meta

Fields

mtPos :: Origin

mtObj :: MetaObj

mtName :: String

mtVal :: String

data MetaObj

Constructors

ContextMeta

data A_Concept

Constructors

PlainConcept

Fields

cptnm :: String

PlainConcept nm represents the set of instances cs by name nm.

ONE
The universal Singleton: I[Anything] = V['Anything'*'Anything']

data A_Gen

data structure A_Gen contains the CLASSIFY statements from an Ampersand script CLASSIFY Employee ISA Person translates to Isa (C Person) (C Employee) CLASSIFY Workingstudent IS Employee/Student translates to IsE orig (C Workingstudent) [C Employee,C Student]

Constructors

Isa

Fields

genspc :: A_Concept

specific concept

gengen :: A_Concept

generic concept

IsE

Fields

genspc :: A_Concept

specific concept

genrhs :: [A_Concept]

concepts of which the conjunction is equivalent to the specific concept

data Sign

Constructors

Sign A_Concept A_Concept

data ConceptDef

Constructors

Cd

Fields

cdpos :: Origin

The position of this definition in the text of the Ampersand source (filename, line number and column number).

cdcpt :: String

The name of the concept for which this is the definition. If there is no such concept, the conceptdefinition is ignored.

cdplug :: Bool

Whether the user specifically told Ampersand not to store this concept in the database

cddef :: String

The textual definition of this concept.

cdtyp :: String

The (SQL) type of this concept.

cdref :: String

A label meant to identify the source of the definition. (useful as LaTeX' symbolic reference)

cdfrom :: String

The name of the pattern or context in which this concept definition was made

class ConceptStructure a where

Minimal complete definition

concs, expressionsIn

Methods

concs

Arguments

:: a

-> [A_Concept]
the set of all concepts used in data structure a

relsUsedIn

Arguments

:: a

-> [Declaration]
the set of all declaratons used within data structure a. `used within` means that there is a relation that refers to that declaration.

relsMentionedIn

Arguments

:: a

-> [Declaration]
the set of all declaratons used within data structure a. `used within` means that there is a relation that refers to that declaration.

primsMentionedIn :: a -> [Expression]

expressionsIn

Arguments

:: a

-> [Expression]
The set of all expressions within data structure a

mp1Pops :: a -> [Population]

mp1Pops draws the population from singleton expressions.

data Activity

Constructors

Act

Fields

actRule :: Rule

actTrig :: [Declaration]

actAffect :: [Declaration]

actQuads :: [Quad]

actEcas :: [ECArule]

actPurp :: [Purpose]

data AMeaning

Constructors

AMeaning

Fields

ameaMrk :: [A_Markup]

data Quad

Constructors

Quad

Fields

qDcl :: Declaration

qRule :: Rule

qConjuncts :: [Conjunct]

data Conjunct

Constructors

Cjct

Fields

rc_id :: String

rc_orgRules :: [Rule]

rc_conjunct :: Expression

rc_dnfClauses :: [DnfClause]

data Fswitchboard

A list of ECA rules, which is used for automated functionality.

Constructors

Fswtch

Fields

fsbEvIn :: [Event]

fsbEvOut :: [Event]

fsbConjs :: [(Rule, Expression)]

fsbECAs :: [ECArule]

data ECArule

Constructors

ECA

Fields

ecaTriggr :: Event

ecaDelta :: Declaration

ecaAction :: PAclause

ecaNum :: Int

data Event

Constructors

On

Fields

eSrt :: InsDel

eDcl :: Declaration

data InsDel

Constructors

Ins

Del

data Pattern

Constructors

A_Pat

Fields

ptnm :: String

Name of this pattern

ptpos :: Origin

the position in the file in which this pattern was declared.

ptend :: Origin

the end position in the file, elements with a position between pos and end are elements of this pattern.

ptrls :: [Rule]

The user defined rules in this pattern

ptgns :: [A_Gen]

The generalizations defined in this pattern

ptdcs :: [Declaration]

The relations that are declared in this pattern

ptups :: [Population]

The user defined populations in this pattern

ptids :: [IdentityDef]

The identity definitions defined in this pattern

ptvds :: [ViewDef]

The view definitions defined in this pattern

ptxps :: [Purpose]

The purposes of elements defined in this pattern

data Declaration

Constructors

Sgn

Fields

decnm :: String

the name of the declaration

decsgn :: Sign

the source and target concepts of the declaration multiplicities returns decprps_calc, when it has been calculated. So if you only need the user defined properties do not use multiplicities but decprps

decprps :: [Prop]

the user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx)

decprps_calc :: Maybe [Prop]

the calculated and user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx, Irf). Note that calculated properties are made by adl2fspec, so in the A-structure decprps and decprps_calc yield exactly the same answer.

decprL :: String

three strings, which form the pragma. E.g. if pragma consists of the three strings: "Person ", " is married to person ", and " in Vegas."

decprM :: String

then a tuple (Peter,Jane) in the list of links means that Person Peter is married to person Jane in Vegas.

decprR :: String

decMean :: AMeaning

the meaning of a declaration, for each language supported by Ampersand.

decfpos :: Origin

the position in the Ampersand source file where this declaration is declared. Not all decalartions come from the ampersand souce file.

deciss :: Bool

if true, this is a signal relation; otherwise it is an ordinary relation.

decusr :: Bool

if true, this relation is declared by an author in the Ampersand script; otherwise it was generated by Ampersand.

decpat :: String

the pattern where this declaration has been declared.

decplug :: Bool

if true, this relation may not be stored in or retrieved from the standard database (it should be gotten from a Plug of some sort instead)

Isn

Fields

detyp :: A_Concept

The type

Vs

Fields

decsgn :: Sign

the source and target concepts of the declaration multiplicities returns decprps_calc, when it has been calculated. So if you only need the user defined properties do not use multiplicities but decprps

data IdentityDef

Constructors

Id

Fields

idPos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

idLbl :: String

the name (or label) of this Identity. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

idCpt :: A_Concept

this expression describes the instances of this object, related to their context

identityAts :: [IdentitySegment]

the constituent attributes (i.e. name/expression pairs) of this identity.

data ViewDef

Constructors

Vd

Fields

vdpos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

vdlbl :: String

the name (or label) of this View. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

vdcpt :: A_Concept

the concept for which this view is applicable

vdIsDefault :: Bool

whether or not this is the default view for the concept

vdhtml :: Maybe ViewHtmlTemplate

the html template for this view (not required since we may have other kinds of views as well in the future) , vdtext :: Maybe ViewText -- Future extension

vdats :: [ViewSegment]

the constituent attributes (i.e. name/expression pairs) of this view.

data IdentitySegment

Constructors

IdentityExp ObjectDef

data ViewSegment

Constructors

ViewExp ObjectDef

ViewText String

ViewHtml String

data Expression

Constructors

EEqu (Expression, Expression)
equivalence =

EImp (Expression, Expression)
implication |-

EIsc (Expression, Expression)
intersection /

EUni (Expression, Expression)
union /

EDif (Expression, Expression)
difference -

ELrs (Expression, Expression)
left residual /

ERrs (Expression, Expression)
right residual

EDia (Expression, Expression)
diamond <>

ECps (Expression, Expression)
composition ;

ERad (Expression, Expression)
relative addition !

EPrd (Expression, Expression)
cartesian product *

EKl0 Expression
Rfx.Trn closure * (Kleene star)

EKl1 Expression
Transitive closure + (Kleene plus)

EFlp Expression
conversion (flip, wok) ~

ECpl Expression
Complement

EBrk Expression
bracketed expression ( ... )

EDcD Declaration
simple declaration

EDcI A_Concept
Identity relation

EEps A_Concept Sign
Epsilon relation (introduced by the system to ensure we compare concepts by equality only.

EDcV Sign
Cartesian product relation

EMp1 String A_Concept
constant (string between single quotes)

data Population

Constructors

PRelPopu

Fields

popdcl :: Declaration

popps :: Pairs

PCptPopu

Fields

popcpt :: A_Concept

popas :: [String]

data FSpec

Constructors

FSpec

Fields

fsName :: String

The name of the specification, taken from the Ampersand script

originalContext :: A_Context

the original context. (for showADL)

getOpts :: Options

The command line options that were used when this FSpec was compiled by Ampersand.

fspos :: [Origin]

The origin of the FSpec. An FSpec can be a merge of a file including other files c.q. a list of Origin.

themes :: [String]

The names of patterns/processes to be printed in the functional specification. (for making partial documentation)

pattsInScope :: [Pattern]

rulesInScope :: [Rule]

declsInScope :: [Declaration]

concsInScope :: [A_Concept]

cDefsInScope :: [ConceptDef]

gensInScope :: [A_Gen]

fsLang :: Lang

The default language for this specification (always specified, so no Maybe here!).

vplugInfos :: [PlugInfo]

All plugs defined in the Ampersand script

plugInfos :: [PlugInfo]

All plugs (defined and derived)

interfaceS :: [Interface]

All interfaces defined in the Ampersand script

interfaceG :: [Interface]

All interfaces derived from the basic ontology (the Lonneker interface)

fSwitchboard :: Fswitchboard

The code to be executed to maintain the truth of invariants

fDeriveProofs :: Blocks

The proofs in Pandoc format

fActivities :: [Activity]

generated: One Activity for every ObjectDef in interfaceG and interfaceS

fRoleRels :: [(Role, Declaration)]

the relation saying which roles may change the population of which relation.

fRoleRuls :: [(Role, Rule)]

the relation saying which roles may change the population of which relation.

fRoles :: [Role]

All roles mentioned in this context.

fallRules :: [Rule]

vrules :: [Rule]

All user defined rules that apply in the entire FSpec

grules :: [Rule]

All rules that are generated: multiplicity rules and identity rules

invariants :: [Rule]

All invariant rules

allUsedDecls :: [Declaration]

All relations that are used in the fSpec

allDecls :: [Declaration]

All relations that are declared in the fSpec

vrels :: [Declaration]

All user defined and generated relations plus all defined and computed totals. The generated relations are all generalizations and one declaration for each signal.

allConcepts :: [A_Concept]

All concepts in the fSpec

kernels :: [[A_Concept]]

All concepts, grouped by their classifications

vIndices :: [IdentityDef]

All keys that apply in the entire FSpec

vviews :: [ViewDef]

All views that apply in the entire FSpec

vgens :: [A_Gen]

All gens that apply in the entire FSpec

vconjs :: [Conjunct]

All conjuncts generated (by ADL2FSpec)

allConjsPerRule :: [(Rule, [Conjunct])]

Maps each rule onto the conjuncts it consists of (note that a single conjunct may be part of several rules)

allConjsPerDecl :: [(Declaration, [Conjunct])]

Maps each declaration to the conjuncts it appears in

allConjsPerConcept :: [(A_Concept, [Conjunct])]

Maps each concept to the conjuncts it appears in (as source or targent of a constituent relation)

vquads :: [Quad]

All quads generated (by ADL2FSpec)

vEcas :: [ECArule]

All ECA rules generated (by ADL2FSpec)

fsisa :: [(A_Concept, A_Concept)]

generated: The data structure containing the generalization structure of concepts

vpatterns :: [Pattern]

All patterns taken from the Ampersand script

conceptDefs :: [ConceptDef]

All concept definitions defined throughout a context, including those inside patterns and processes

fSexpls :: [Purpose]

All purposes that have been declared at the top level of the current specification, but not in the processes, patterns and interfaces.

metas :: [Meta]

All meta relations from the entire context

crudInfo :: CrudInfo

Information for CRUD matrices

initialPops :: [Population]

All user defined populations of relations and concepts

allAtoms :: [Atom]

allLinks :: [A_Pair]

initialConjunctSignals :: [(Conjunct, [Paire])]

All conjuncts that have process-rule violations.

allViolations :: [(Rule, [Paire])]

All invariant rules with violations.

allExprs :: [Expression]

All expressions in the fSpec

allSigns :: [Sign]

All Signs in the fSpec

concDefs :: FSpec -> A_Concept -> [ConceptDef]

data PlugSQL

Constructors

TblSQL
stores a related collection of relations: a kernel of concepts and attribute relations of this kernel i.e. a list of SqlField given some A -> [target r | r::A*B,isUni r,isTot r, isInj r] ++ [target r | r::A*B,isUni r, not(isTot r), not(isSur r)] kernel = A closure of concepts A,B for which there exists a r::A->B[INJ] (r=fldexpr of kernel field holding instances of B, in practice r is I or a makeRelation(flipped declaration)) attribute relations = All concepts B, A in kernel for which there exists a r::A*B[UNI] and r not TOT and SUR (r=fldexpr of attMor field, in practice r is a makeRelation(declaration))

Fields

sqlname :: String

fields :: [SqlField]

the first field is the concept table of the most general concept (e.g. Person) then follow concept tables of specializations. Together with the first field this is called the "kernel" the remaining fields represent attributes.

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkpTbl :: [(Expression, SqlField, SqlField)]

lookup table that links concepts to column names in the plug (kernel+attRels) mLkpTbl is een lijst met relaties die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

BinSQL
stores one relation r in two ordered columns i.e. a tuple of SqlField -> (source r,target r) with (fldexpr=I/r;r~, fldexpr=r) (note: if r TOT then (I/r;r~ = I). Thus, the concept (source r) is stored in this plug too) with tblcontents = [[Just x,Just y] |(x,y)<-contents r]. Typical for BinSQL is that it has exactly two columns that are not unique and may not contain NULL values

Fields

sqlname :: String

columns :: (SqlField, SqlField)

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkp :: Expression

ScalarSQL
stores one concept c in one column i.e. a SqlField -> c with tblcontents = [[Just x] |(x,_)<-contents c]. Typical for ScalarSQL is that it has exactly one column that is unique and may not contain NULL values i.e. fldexpr=I[c]

Fields

sqlname :: String

sqlColumn :: SqlField

cLkp :: A_Concept

data SqlField

Constructors

Fld

Fields

fldname :: String

fldexpr :: Expression

De target van de expressie geeft de waarden weer in de SQL-tabel-kolom.

fldtype :: SqlType

flduse :: SqlFieldUsage

fldnull :: Bool

True if there can be empty field-values (intended for data dictionary of DB-implementation)

flduniq :: Bool

True if all field-values are unique? (intended for data dictionary of DB-implementation)

data SqlType

Constructors

SQLChar Int

SQLBlob

SQLPass

SQLSingle

SQLDouble

SQLText

SQLuInt Int

SQLsInt Int

SQLId

SQLVarchar Int

SQLBool

data PlugInfo

Constructors

InternalPlug PlugSQL

ExternalPlug ObjectDef

data PAclause

Constructors

CHC

Fields

paCls :: [PAclause]

paMotiv :: [(Expression, [Rule])]

GCH

Fields

paGCls :: [(InsDel, Expression, PAclause)]

paMotiv :: [(Expression, [Rule])]

ALL

Fields

paCls :: [PAclause]

paMotiv :: [(Expression, [Rule])]

Do

Fields

paSrt :: InsDel

paTo :: Declaration

paDelta :: Expression

paMotiv :: [(Expression, [Rule])]

New

Fields

paCpt :: A_Concept

paCl :: String -> PAclause

paMotiv :: [(Expression, [Rule])]

Rmv

Fields

paCpt :: A_Concept

paCl :: String -> PAclause

paMotiv :: [(Expression, [Rule])]

Nop

Fields

paMotiv :: [(Expression, [Rule])]

Blk

Fields

paMotiv :: [(Expression, [Rule])]

Let

Fields

paExpr :: PAclause

paBody :: PAclause -> PAclause

paMotiv :: [(Expression, [Rule])]

Ref

Fields

paVar :: String

data Rule

Constructors

Ru

Fields

rrnm :: String

Name of this rule

rrexp :: Expression

The rule expression

rrfps :: Origin

Position in the Ampersand file

rrmean :: AMeaning

Ampersand generated meaning (for all known languages)

rrmsg :: [A_Markup]

User-specified violation messages, possibly more than one, for multiple languages.

rrviol :: Maybe (PairView Expression)

Custom presentation for violations, currently only in a single language

rrtyp :: Sign

Allocated type

rrdcl :: Maybe (Prop, Declaration)

The property, if this rule originates from a property on a Declaration

r_env :: String

Name of pattern in which it was defined.

r_usr :: RuleOrigin

Where does this rule come from?

isSignal :: Bool

True if this is a signal; False if it is an invariant

data Prop

Constructors

Uni
univalent

Inj
injective

Sur
surjective

Tot
total

Sym
symmetric

Asy
antisymmetric

Trn
transitive

Rfx
reflexive

Irf
irreflexive

Aut
automatically computed (NOTE: this is a hacky way to denote these until we have appropriate syntax)

Prop
PROP keyword, later replaced by [Sym, Asy]

data RuleOrigin

Constructors

UserDefined

Multiplicity

Identity

data Lang

Constructors

Dutch

English

data SqlFieldUsage

Constructors

TableKey Bool A_Concept

ForeignKey A_Concept

PlainAttr

data DnfClause

Constructors

Dnf

Fields

antcs :: [Expression]

conss :: [Expression]

data Options

This data constructor is able to hold all kind of information that is useful to express what the user would like Ampersand to do.

Constructors

Options

Fields

showVersion :: Bool

preVersion :: String

postVersion :: String

showHelp :: Bool

verboseP :: Bool

development :: Bool

validateSQL :: Bool

validateEdit :: Maybe String

genPrototype :: Bool

dirPrototype :: String

allInterfaces :: Bool

dbName :: String

genAtlas :: Bool

namespace :: String

autoRefresh :: Maybe Int

testRule :: Maybe String

customCssFile :: Maybe FilePath

importfile :: FilePath

fileformat :: Maybe FileFormat

theme :: DocTheme

genXML :: Bool

genFSpec :: Bool

diag :: Bool

fspecFormat :: FSpecFormat

genGraphics :: Bool

genEcaDoc :: Bool

proofs :: Bool

haskell :: Bool

dirOutput :: String

outputfile :: String

crowfoot :: Bool

blackWhite :: Bool

doubleEdges :: Bool

showPredExpr :: Bool

noDiagnosis :: Bool

diagnosisOnly :: Bool

genLegalRefs :: Bool

genUML :: Bool

genFPAChap :: Bool

genFPAExcel :: Bool

genStaticFiles :: Bool

genBericht :: Bool

language :: Maybe Lang

dirExec :: String

ampersandDataDir :: FilePath

progrName :: String

fileName :: FilePath

baseName :: String

logName :: FilePath

genTime :: LocalTime

export2adl :: Bool

test :: Bool

genASTTables :: Bool

genASTFile :: Bool

genGenericsFile :: Bool

genGenericTables :: Bool

metaTablesHaveUnderscore :: Bool

pangoFont :: String

sqlHost :: String

sqlLogin :: String

sqlPwd :: String

oldNormalizer :: Bool

newFrontend :: Bool

data DocTheme

Constructors

DefaultTheme

ProofTheme

StudentTheme

StudentDesignerTheme

DesignerTheme

data Picture

Constructors

Pict

Fields

pType :: PictureReq

the type of the picture

scale :: String

a scale factor, intended to pass on to LaTeX, because Pandoc seems to have a problem with scaling.

dotSource :: DotGraph String

the string representing the .dot

dotProgName :: GraphvizCommand

the name of the program to use ("dot" or "neato" or "fdp" or Sfdp)

caption :: String

a human readable name of this picture

writePicture :: Options -> Picture -> IO ()

data PictureReq

Constructors

PTClassDiagram

PTRelsUsedInPat Pattern

PTDeclaredInPat Pattern

PTConcept A_Concept

PTSwitchBoard Activity

PTFinterface Activity

PTIsaInPattern Pattern

PTSingleRule Rule

PTLogicalDM

PTTechnicalDM

makePicture :: FSpec -> PictureReq -> Picture

data FilePos

Represents a position within a file, including the file path, line and column numbers

Constructors

FilePos FilePath Line Column

data Origin

Constructors

OriginUnknown

Origin String

FileLoc FilePos SymbolName

DBLoc String

mkPair :: String -> String -> Paire

Classes:

class Association rel where

Minimal complete definition

sign

Methods

source, target :: rel -> A_Concept

sign :: rel -> Sign

isEndo :: rel -> Bool

flp :: Flippable a => a -> a

class Collection a where

Methods

eleM :: Eq b => b -> a b -> Bool

uni, isc :: Eq b => a b -> a b -> a b

(>-) :: Eq b => a b -> a b -> a b infixl 5

empty :: Eq b => a b

elems :: Eq b => a b -> [b]

class Named a where

anything could have some label, can't it?

Methods

name :: a -> String

data ObjectDef

Constructors

Obj

Fields

objnm :: String

view name of the object definition. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface if it is not an empty string.

objpos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number)

objctx :: Expression

this expression describes the instances of this object, related to their context.

objmView :: Maybe String

The view that should be used for this object

objmsub :: Maybe SubInterface

the attributes, which are object definitions themselves.

objstrs :: [[String]]

directives that specify the interface.

class Association r => Relational r where

Minimal complete definition

multiplicities, isProp, isImin, isTrue, isFalse, isIdent, isEpsilon

Methods

multiplicities :: r -> [Prop]

isProp :: r -> Bool

isImin :: r -> Bool

isTrue :: r -> Bool

isFalse :: r -> Bool

isFunction :: r -> Bool

isTot :: r -> Bool

isUni :: r -> Bool

isSur :: r -> Bool

isInj :: r -> Bool

isRfx :: r -> Bool

isIrf :: r -> Bool

isTrn :: r -> Bool

isSym :: r -> Bool

isAsy :: r -> Bool

isIdent :: r -> Bool

isEpsilon :: r -> Bool

data Interface

Constructors

Ifc

Fields

ifcParams :: [Declaration]

ifcClass :: Maybe String

ifcArgs :: [[String]]

ifcRoles :: [Role]

ifcObj :: ObjectDef

ifcEcas :: [ECArule]

ifcControls :: [Conjunct]

ifcPos :: Origin

ifcPrp :: String

getInterfaceByName :: [Interface] -> String -> Interface

data SubInterface

Constructors

Box A_Concept (Maybe String) [ObjectDef]

InterfaceRef String

class Object a where

Methods

concept :: a -> A_Concept

attributes :: a -> [ObjectDef]

contextOf :: a -> Expression

class (Named p, Eq p, Show p) => Plugable p where

Methods

makePlug :: PlugInfo -> p

class Named a => Motivated a where

Minimal complete definition

explForObj, explanations

Methods

purposeOf

Arguments

:: FSpec

-> Lang

-> a

-> Maybe [Purpose]
explains the purpose of a, i.e. the reason why a exists. The purpose could be either given by the user, or generated by Ampersand. Multiple purposes are allowed for the following reasons: * Different purposes from different sources make me want to document them all. * Combining two overlapping scripts from (i.e. from different authors) may cause multiple purposes.

explForObj

Arguments

:: a

-> ExplObj

-> Bool
Given an Explainable object and an ExplObj, return TRUE if they concern the identical object.

explanations

Arguments

:: a

-> [Purpose]
The explanations that are defined inside a (including that of a itself)

purposesDefinedIn

Arguments

:: FSpec

-> Lang

-> a

-> [Purpose]
The explanations that are defined inside a (including that of a itself)

class Traced a where

Minimal complete definition

origin

Methods

origin :: a -> Origin

filenm :: a -> String

linenr :: a -> Int

colnr :: a -> Int

class Language a where

Minimal complete definition

relsDefdIn, udefrules, identities, viewDefs, gens, patterns

Methods

relsDefdIn

Arguments

:: a

-> [Declaration]
all relations that are declared in the scope of this viewpoint. These are user defined relations and all generated relarations, i.e. one relation for each GEN and one for each signal rule. Don't confuse relsDefdIn with relsUsedIn, which gives the relations that are used in a.)

udefrules

Arguments

:: a

-> [Rule]
all user defined rules that are maintained within this viewpoint, which are not multiplicity- and not identity rules.

multrules

Arguments

:: a

-> [Rule]
all multiplicityrules that are maintained within this viewpoint.

identityRules :: a -> [Rule]

allRules :: a -> [Rule]

identities

Arguments

:: a

-> [IdentityDef]
all keys that are defined in a

viewDefs

Arguments

:: a

-> [ViewDef]
all views that are defined in a

gens

Arguments

:: a

-> [A_Gen]
all generalizations that are valid within this viewpoint

patterns

Arguments

:: a

-> [Pattern]
all patterns that are used in this viewpoint

class ShowHS a where

Methods

showHS :: Options -> String -> a -> String

class ShowHSName a where

Methods

showHSName :: a -> String

haskellIdentifier :: String -> String

Functions on concepts

(<==>) :: Poset a => a -> a -> Bool

Is comparable to.

meet :: Sortable a => a -> a -> a

join :: Sortable a => a -> a -> a

sortWith :: (Show b, Poset b) => (b -> [[b]], [b]) -> (a -> b) -> [a] -> [a]

atomsOf :: [A_Gen] -> [Population] -> A_Concept -> [String]

This function returns the atoms of a concept (like fullContents does for relation-like things.)

smallerConcepts :: [A_Gen] -> A_Concept -> [A_Concept]

this function takes all generalisation relations from the context and a concept and delivers a list of all concepts that are more specific than the given concept. If there are no cycles in the generalization graph, cpt cannot be an element of smallerConcepts gens cpt.

largerConcepts :: [A_Gen] -> A_Concept -> [A_Concept]

this function takes all generalisation relations from the context and a concept and delivers a list of all concepts that are more generic than the given concept.

rootConcepts :: [A_Gen] -> [A_Concept] -> [A_Concept]

this function returns the most generic concepts in the class of a given concept

Functions on relations

Functions on rules

Functions on expressions:

conjNF :: Options -> Expression -> Expression

disjNF :: Options -> Expression -> Expression

simplify :: Expression -> Expression

cfProof :: Options -> Expression -> Proof Expression

dfProof :: Options -> Expression -> Proof Expression

normPA :: Options -> PAclause -> PAclause

lookupCpt :: FSpec -> A_Concept -> [(PlugSQL, SqlField)]

This returns all columntable pairs that serve as a concept table for cpt. When addingremoving atoms, all of these columns need to be updated

showPrf :: (expr -> String) -> Proof expr -> [String]

notCpl :: Expression -> Expression

isCpl :: Expression -> Bool

isPos :: Expression -> Bool

isNeg :: Expression -> Bool

foldrMapExpression :: (r -> a -> a) -> (Declaration -> r) -> a -> Expression -> a

(.==.) :: Expression -> Expression -> Expression infixl 1

(.|-.) :: Expression -> Expression -> Expression infixl 1

(./\.) :: Expression -> Expression -> Expression infixl 2

(.\/.) :: Expression -> Expression -> Expression infixl 2

(.-.) :: Expression -> Expression -> Expression infixl 4

(./.) :: Expression -> Expression -> Expression infixl 6

(.\.) :: Expression -> Expression -> Expression infixl 6

(.<>.) :: Expression -> Expression -> Expression infixl 6

(.:.) :: Expression -> Expression -> Expression infixl 8

(.!.) :: Expression -> Expression -> Expression infixl 8

(.*.) :: Expression -> Expression -> Expression infixl 8

deMorganERad :: Expression -> Expression

The rule of De Morgan requires care with respect to the complement. The following function provides a function to manipulate with De Morgan correctly.

deMorganECps :: Expression -> Expression

deMorganEUni :: Expression -> Expression

deMorganEIsc :: Expression -> Expression

exprUni2list :: Expression -> [Expression]

exprIsc2list :: Expression -> [Expression]

exprCps2list :: Expression -> [Expression]

exprRad2list :: Expression -> [Expression]

exprPrd2list :: Expression -> [Expression]

Functions with plugs:

showPlug :: PlugSQL -> [String]

plugFields :: PlugSQL -> [SqlField]

tblcontents :: [A_Gen] -> [Population] -> PlugSQL -> [TblRecord]

fldauto :: SqlField -> Bool

Parser related stuff

parseADL1pExpr

Arguments

:: String
The string to be parsed

-> FilePath
The name of the file (used for error messages)

-> Either String (Term TermPrim)
The result: Either an error message, or a good result

Parses an isolated ADL1 expression string

data CtxError

createFSpec

Arguments

:: Options
The options derived from the command line

-> IO (Guarded FSpec)

create an FSpec, based on the provided command-line options.

getGeneralizations :: FSpec -> A_Concept -> [A_Concept]

getSpecializations :: FSpec -> A_Concept -> [A_Concept]

Type checking and calculus

data Guarded a

Constructors

Errors [CtxError]

Checked a

pCtx2aCtx :: Options -> P_Context -> Guarded A_Context

makeFSpec :: Options -> A_Context -> FSpec

getPopulationsFrom :: Options -> FilePath -> IO (Guarded [Population])

Generators of output

generateAmpersandOutput :: FSpec -> IO ()

Prettyprinters

class ShowADL a where

Methods

showADL :: a -> String

showSQL :: SqlType -> String

showSign :: Association a => a -> String

Functions with Options

getOptions :: IO Options

verboseLn :: Options -> String -> IO ()

verbose :: Options -> String -> IO ()

data FileFormat

Constructors

Adl1Format

Adl1PopFormat

helpNVersionTexts :: String -> Options -> [String]

Other functions

eqCl :: Eq b => (a -> b) -> [a] -> [[a]]

eqCl is used for gathering things that are equal wrt some criterion f. For instance, if you want to have persons with the same name: 'eqCl name persons' produces a list,in which each element is a list of persons with the same name. Example> eqCl (==s) Mississippi = "ssss"

showErr :: CtxError -> String

unCap :: String -> String

Converts the first character of a string to lowercase, with the exception that there is a second character, which is uppercase. uncap AbcDe == "abcDe" uncap ABcDE == ABcDE

upCap :: String -> String

Converts the first character of a string to uppercase

escapeNonAlphaNum :: String -> String

escape anything except regular characters and digits to _code e.g. escapeNonAlphaNum "a_é" = "a_95_233"

fatalMsg :: String -> Int -> String -> a

a function to create error message in a structured way, containing the version of Ampersand. It throws an error, showing a (module)name and a number. This makes debugging pretty easy.

ampersandVersionStr :: String

String, containing the Ampersand version, including the build timestamp.

ampersandVersionWithoutBuildTimeStr :: String

String, containing the Ampersand version

putStr :: String -> IO ()

hGetContents :: Handle -> IO String

hPutStr :: Handle -> String -> IO ()

hPutStrLn :: Handle -> String -> IO ()

readFile :: FilePath -> IO String

writeFile :: FilePath -> String -> IO ()

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thd3 :: (a, b, c) -> c

blockParenthesize :: String -> String -> String -> [[String]] -> [String]

addToLastLine :: String -> [String] -> [String]

indent :: Int -> [String] -> [String]

trace :: String -> a -> a

The trace function outputs the trace message given as its first argument, before returning the second argument as its result.

For example, this returns the value of f x but first outputs the message.

trace ("calling f with x = " ++ show x) (f x)

The trace function should only be used for debugging, or for monitoring execution. The function is not referentially transparent: its type indicates that it is a pure function but it has the side effect of outputting the trace message.

showTrace :: Show a => a -> a

showTraceTag :: Show a => String -> a -> a

Stuff that should probably not be in the prototype

data A_Markup

Constructors

A_Markup

Fields

amLang :: Lang

amFormat :: PandocFormat

amPandoc :: [Block]

blocks2String :: PandocFormat -> Bool -> [Block] -> String

write [Block] as String in a certain format using defaultWriterOptions

aMarkup2String :: A_Markup -> String

data PandocFormat

Constructors

HTML

ReST

LaTeX

Markdown

class Meaning a where

Minimal complete definition

meaning

Methods

meaning :: Lang -> a -> Maybe A_Markup

meaning2Blocks :: Lang -> a -> [Block]

rulefromProp :: Prop -> Declaration -> Maybe Rule

fullContents :: [A_Gen] -> [Population] -> Expression -> Pairs

data Paire

srcPaire :: Paire -> String

trgPaire :: Paire -> String

data Purpose

Explanation is the intended constructor. It explains the purpose of the object it references. The enrichment process of the parser must map the names (from PPurpose) to the actual objects

Constructors

Expl

Fields

explPos :: Origin

The position in the Ampersand script of this purpose definition

explObj :: ExplObj

The object that is explained.

explMarkup :: A_Markup

This field contains the text of the explanation including language and markup info.

explUserdefd :: Bool

Is this purpose defined in the script?

explRefIds :: [String]

The references of the explaination

data ExplObj

Constructors

ExplConceptDef ConceptDef

ExplDeclaration Declaration

ExplRule String

ExplIdentityDef String

ExplViewDef String

ExplPattern String

ExplInterface String

ExplContext String