Database.Design.Ampersand.ADL1.Expression

Documentation

subst :: (Declaration, Expression) -> Expression -> Expression

subst is used to replace each occurrence of a relation with an expression. The parameter expr will therefore be applied to an expression of the form Erel rel.

foldlMapExpression :: (a -> r -> a) -> (Declaration -> r) -> a -> Expression -> a

foldrMapExpression :: (r -> a -> a) -> (Declaration -> r) -> a -> Expression -> a

primitives :: Expression -> [Expression]

isMp1 :: Expression -> Bool

isEEps :: Expression -> Bool

isPos :: Expression -> Bool

isNeg :: Expression -> Bool

deMorganERad :: Expression -> Expression

The rule of De Morgan requires care with respect to the complement. The following function provides a function to manipulate with De Morgan correctly.

deMorganECps :: Expression -> Expression

deMorganEUni :: Expression -> Expression

deMorganEIsc :: Expression -> Expression

notCpl :: Expression -> Expression

isCpl :: Expression -> Bool

exprIsc2list :: Expression -> [Expression]

exprUni2list :: Expression -> [Expression]

exprCps2list :: Expression -> [Expression]

exprRad2list :: Expression -> [Expression]

exprPrd2list :: Expression -> [Expression]

insParentheses :: Expression -> Expression