Database.Design.Ampersand.ADL1.Lattices

Documentation

findExact :: (Ord a, SetLike x) => Op1EqualitySystem a -> FreeLattice a -> x a

findSubsets :: (Ord a, SetLike x) => Op1EqualitySystem a -> FreeLattice a -> [x a]

optimize1 :: Ord a => EqualitySystem a -> Op1EqualitySystem a

data Op1EqualitySystem a

addEquality :: (Ord a, SetLike x) => (x a, x a) -> EqualitySystem a -> EqualitySystem a

emptySystem :: EqualitySystem a

data FreeLattice a

Constructors

Join (FreeLattice a) (FreeLattice a)

Meet (FreeLattice a) (FreeLattice a)

Atom a

getGroups :: (Ord concept, SetLike set) => Op1EqualitySystem concept -> [set concept]

getGroups: create groups of concepts (type variable: concept). 1. Each concept is in precisely one group. 2. Two concepts are in the same group if there is a path of classify-rules between those concepts. The purpose of this is to know whether two concepts are comparable or not. Atoms of concepts within a group can be compared. Atoms of concepts in different groups may never be compared.

isInSystem :: Ord a => Op1EqualitySystem a -> a -> Bool