Database.Design.Ampersand.Basics.Auxiliaries

Documentation

eqClass :: (a -> a -> Bool) -> [a] -> [[a]]

The eqClass function takes an equality test function and a list and returns a list of lists such that each sublist in the result contains only equal elements, and all equal elements are in the same sublist. For example,

Example> eqClass Mississippi = [M,"iiii","ssss","pp"]

eqCl :: Eq b => (a -> b) -> [a] -> [[a]]

eqCl is used for gathering things that are equal wrt some criterion f. For instance, if you want to have persons with the same name: 'eqCl name persons' produces a list,in which each element is a list of persons with the same name. Example> eqCl (==s) Mississippi = "ssss"

getCycles :: Eq a => [(a, [a])] -> [[a]]

getCycles returns a list of cycles in the edges list (each edge is a pair of a from-vertex and a list of to-vertices)

transClosureMap :: (Eq a, Ord a) => Map a [a] -> Map a [a]

Warshall's transitive closure algorithm

combinations :: [[a]] -> [[a]]

The combinations function returns all possible combinations of lists of list. For example,

combinations [[1,2,3],[10,20],[4]] == [[1,10,4],[1,20,4],[2,10,4],[2,20,4],[3,10,4],[3,20,4]]

converse :: forall a b. (Ord a, Ord b) => [(a, [b])] -> [(b, [a])]

commaEng :: String -> [String] -> String

commaNL :: String -> [String] -> String

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thd3 :: (a, b, c) -> c

class Flippable a where

Methods

flp :: a -> a

showTrace :: Show a => a -> a

showTraceTag :: Show a => String -> a -> a

blockParenthesize :: String -> String -> String -> [[String]] -> [String]

addToLastLine :: String -> [String] -> [String]

indent :: Int -> [String] -> [String]

module Debug.Trace