Database.Design.Ampersand.Basics.String

Description

This module contains some common String funcions

Documentation

unCap :: String -> String

Converts the first character of a string to lowercase, with the exception that there is a second character, which is uppercase. uncap AbcDe == "abcDe" uncap ABcDE == ABcDE

upCap :: String -> String

Converts the first character of a string to uppercase

escapeNonAlphaNum :: String -> String

escape anything except regular characters and digits to _code e.g. escapeNonAlphaNum "a_é" = "a_95_233"