Database.Design.Ampersand.Basics.Version

Description

This module contains Version of Ampersand

Documentation

ampersandVersionStr :: String

String, containing the Ampersand version, including the build timestamp.

ampersandVersionWithoutBuildTimeStr :: String

String, containing the Ampersand version

fatalMsg :: String -> Int -> String -> a

a function to create error message in a structured way, containing the version of Ampersand. It throws an error, showing a (module)name and a number. This makes debugging pretty easy.