Database.Design.Ampersand.Classes.ConceptStructure

Documentation

class ConceptStructure a where

Minimal complete definition

concs, expressionsIn

Methods

concs

Arguments

:: a

-> [A_Concept]
the set of all concepts used in data structure a

relsUsedIn

Arguments

:: a

-> [Declaration]
the set of all declaratons used within data structure a. `used within` means that there is a relation that refers to that declaration.

relsMentionedIn

Arguments

:: a

-> [Declaration]
the set of all declaratons used within data structure a. `used within` means that there is a relation that refers to that declaration.

primsMentionedIn :: a -> [Expression]

expressionsIn

Arguments

:: a

-> [Expression]
The set of all expressions within data structure a

mp1Pops :: a -> [Population]

mp1Pops draws the population from singleton expressions.