Database.Design.Ampersand.Classes.ViewPoint

Documentation

class Language a where

Minimal complete definition

relsDefdIn, udefrules, identities, viewDefs, gens, patterns

Methods

relsDefdIn

Arguments

:: a

-> [Declaration]
all relations that are declared in the scope of this viewpoint. These are user defined relations and all generated relarations, i.e. one relation for each GEN and one for each signal rule. Don't confuse relsDefdIn with relsUsedIn, which gives the relations that are used in a.)

udefrules

Arguments

:: a

-> [Rule]
all user defined rules that are maintained within this viewpoint, which are not multiplicity- and not identity rules.

multrules

Arguments

:: a

-> [Rule]
all multiplicityrules that are maintained within this viewpoint.

identityRules :: a -> [Rule]

allRules :: a -> [Rule]

identities

Arguments

:: a

-> [IdentityDef]
all keys that are defined in a

viewDefs

Arguments

:: a

-> [ViewDef]
all views that are defined in a

gens

Arguments

:: a

-> [A_Gen]
all generalizations that are valid within this viewpoint

patterns

Arguments

:: a

-> [Pattern]
all patterns that are used in this viewpoint