Database.Design.Ampersand.Components

Description

This module contains the building blocks that are available in the Ampersand Library. These building blocks will be described further at [ampersand.sourceforge.net |the wiki pages of our project].

Type checking and calculus

makeFSpec :: Options -> A_Context -> FSpec

Generators of output

generateAmpersandOutput :: FSpec -> IO ()

etc...