Database.Design.Ampersand.Core.AbstractSyntaxTree

Documentation

data A_Context

Constructors

ACtx

Fields

ctxnm :: String

The name of this context

ctxpos :: [Origin]

The origin of the context. A context can be a merge of a file including other files c.q. a list of Origin.

ctxlang :: Lang

The default language used in this context.

ctxmarkup :: PandocFormat

The default markup format for free text in this context (currently: LaTeX, ...)

ctxthms :: [String]

Names of patterns/processes to be printed in the functional specification. (For partial documents.)

ctxpats :: [Pattern]

The patterns defined in this context

ctxrs :: [Rule]

All user defined rules in this context, but outside patterns and outside processes

ctxds :: [Declaration]

The relations that are declared in this context, outside the scope of patterns

ctxpopus :: [Population]

The user defined populations of relations defined in this context, including those from patterns and processes

ctxcds :: [ConceptDef]

The concept definitions defined in this context, including those from patterns and processes

ctxks :: [IdentityDef]

The identity definitions defined in this context, outside the scope of patterns

ctxrrules :: [A_RoleRule]

ctxRRels :: [A_RoleRelation]

The assignment of roles to Relations (which role mayEdit what relations)

ctxvs :: [ViewDef]

The view definitions defined in this context, outside the scope of patterns

ctxgs :: [A_Gen]

The specialization statements defined in this context, outside the scope of patterns

ctxgenconcs :: [[A_Concept]]

A partitioning of all concepts: the union of all these concepts contains all atoms, and the concept-lists are mutually distinct in terms of atoms in one of the mentioned concepts

ctxifcs :: [Interface]

The interfaces defined in this context

ctxps :: [Purpose]

The purposes of objects defined in this context, outside the scope of patterns and processes

ctxsql :: [ObjectDef]

user defined sqlplugs, taken from the Ampersand script

ctxphp :: [ObjectDef]

user defined phpplugs, taken from the Ampersand script

ctxmetas :: [Meta]

used for Pandoc authors (and possibly other things)

data Meta

Constructors

Meta

Fields

mtPos :: Origin

mtObj :: MetaObj

mtName :: String

mtVal :: String

data Pattern

Constructors

A_Pat

Fields

ptnm :: String

Name of this pattern

ptpos :: Origin

the position in the file in which this pattern was declared.

ptend :: Origin

the end position in the file, elements with a position between pos and end are elements of this pattern.

ptrls :: [Rule]

The user defined rules in this pattern

ptgns :: [A_Gen]

The generalizations defined in this pattern

ptdcs :: [Declaration]

The relations that are declared in this pattern

ptups :: [Population]

The user defined populations in this pattern

ptids :: [IdentityDef]

The identity definitions defined in this pattern

ptvds :: [ViewDef]

The view definitions defined in this pattern

ptxps :: [Purpose]

The purposes of elements defined in this pattern

data PairView a

Constructors

PairView

Fields

ppv_segs :: [PairViewSegment a]

data PairViewSegment a

Constructors

PairViewText

Fields

pvsOrg :: Origin

pvsStr :: String

PairViewExp

Fields

pvsOrg :: Origin

pvsSoT :: SrcOrTgt

pvsExp :: a

data Rule

Constructors

Ru

Fields

rrnm :: String

Name of this rule

rrexp :: Expression

The rule expression

rrfps :: Origin

Position in the Ampersand file

rrmean :: AMeaning

Ampersand generated meaning (for all known languages)

rrmsg :: [A_Markup]

User-specified violation messages, possibly more than one, for multiple languages.

rrviol :: Maybe (PairView Expression)

Custom presentation for violations, currently only in a single language

rrtyp :: Sign

Allocated type

rrdcl :: Maybe (Prop, Declaration)

The property, if this rule originates from a property on a Declaration

r_env :: String

Name of pattern in which it was defined.

r_usr :: RuleOrigin

Where does this rule come from?

isSignal :: Bool

True if this is a signal; False if it is an invariant

ruleIsInvariantUniOrInj :: Rule -> Bool

data RuleType

Constructors

Implication

Equivalence

Truth

data RuleOrigin

Constructors

UserDefined

Multiplicity

Identity

data Declaration

Constructors

Sgn

Fields

decnm :: String

the name of the declaration

decsgn :: Sign

the source and target concepts of the declaration multiplicities returns decprps_calc, when it has been calculated. So if you only need the user defined properties do not use multiplicities but decprps

decprps :: [Prop]

the user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx)

decprps_calc :: Maybe [Prop]

the calculated and user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx, Irf). Note that calculated properties are made by adl2fspec, so in the A-structure decprps and decprps_calc yield exactly the same answer.

decprL :: String

three strings, which form the pragma. E.g. if pragma consists of the three strings: "Person ", " is married to person ", and " in Vegas."

decprM :: String

then a tuple (Peter,Jane) in the list of links means that Person Peter is married to person Jane in Vegas.

decprR :: String

decMean :: AMeaning

the meaning of a declaration, for each language supported by Ampersand.

decfpos :: Origin

the position in the Ampersand source file where this declaration is declared. Not all decalartions come from the ampersand souce file.

deciss :: Bool

if true, this is a signal relation; otherwise it is an ordinary relation.

decusr :: Bool

if true, this relation is declared by an author in the Ampersand script; otherwise it was generated by Ampersand.

decpat :: String

the pattern where this declaration has been declared.

decplug :: Bool

if true, this relation may not be stored in or retrieved from the standard database (it should be gotten from a Plug of some sort instead)

Isn

Fields

detyp :: A_Concept

The type

Vs

Fields

decsgn :: Sign

the source and target concepts of the declaration multiplicities returns decprps_calc, when it has been calculated. So if you only need the user defined properties do not use multiplicities but decprps

data IdentityDef

Constructors

Id

Fields

idPos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

idLbl :: String

the name (or label) of this Identity. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

idCpt :: A_Concept

this expression describes the instances of this object, related to their context

identityAts :: [IdentitySegment]

the constituent attributes (i.e. name/expression pairs) of this identity.

data IdentitySegment

Constructors

IdentityExp ObjectDef

data ViewDef

Constructors

Vd

Fields

vdpos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

vdlbl :: String

the name (or label) of this View. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

vdcpt :: A_Concept

the concept for which this view is applicable

vdIsDefault :: Bool

whether or not this is the default view for the concept

vdhtml :: Maybe ViewHtmlTemplate

the html template for this view (not required since we may have other kinds of views as well in the future) , vdtext :: Maybe ViewText -- Future extension

vdats :: [ViewSegment]

the constituent attributes (i.e. name/expression pairs) of this view.

data ViewSegment

Constructors

ViewExp ObjectDef

ViewText String

ViewHtml String

data A_Gen

data structure A_Gen contains the CLASSIFY statements from an Ampersand script CLASSIFY Employee ISA Person translates to Isa (C Person) (C Employee) CLASSIFY Workingstudent IS Employee/Student translates to IsE orig (C Workingstudent) [C Employee,C Student]

Constructors

Isa

Fields

genspc :: A_Concept

specific concept

gengen :: A_Concept

generic concept

IsE

Fields

genspc :: A_Concept

specific concept

genrhs :: [A_Concept]

concepts of which the conjunction is equivalent to the specific concept

data Interface

Constructors

Ifc

Fields

ifcParams :: [Declaration]

ifcClass :: Maybe String

ifcArgs :: [[String]]

ifcRoles :: [Role]

ifcObj :: ObjectDef

ifcEcas :: [ECArule]

ifcControls :: [Conjunct]

ifcPos :: Origin

ifcPrp :: String

getInterfaceByName :: [Interface] -> String -> Interface

data SubInterface

Constructors

Box A_Concept (Maybe String) [ObjectDef]

InterfaceRef String

data ObjectDef

Constructors

Obj

Fields

objnm :: String

view name of the object definition. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface if it is not an empty string.

objpos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number)

objctx :: Expression

this expression describes the instances of this object, related to their context.

objmView :: Maybe String

The view that should be used for this object

objmsub :: Maybe SubInterface

the attributes, which are object definitions themselves.

objstrs :: [[String]]

directives that specify the interface.

class Object a where

Methods

concept :: a -> A_Concept

attributes :: a -> [ObjectDef]

contextOf :: a -> Expression

objAts :: ObjectDef -> [ObjectDef]

data Purpose

Explanation is the intended constructor. It explains the purpose of the object it references. The enrichment process of the parser must map the names (from PPurpose) to the actual objects

Constructors

Expl

Fields

explPos :: Origin

The position in the Ampersand script of this purpose definition

explObj :: ExplObj

The object that is explained.

explMarkup :: A_Markup

This field contains the text of the explanation including language and markup info.

explUserdefd :: Bool

Is this purpose defined in the script?

explRefIds :: [String]

The references of the explaination

data ExplObj

Constructors

ExplConceptDef ConceptDef

ExplDeclaration Declaration

ExplRule String

ExplIdentityDef String

ExplViewDef String

ExplPattern String

ExplInterface String

ExplContext String

data Expression

Constructors

EEqu (Expression, Expression)
equivalence =

EImp (Expression, Expression)
implication |-

EIsc (Expression, Expression)
intersection /

EUni (Expression, Expression)
union /

EDif (Expression, Expression)
difference -

ELrs (Expression, Expression)
left residual /

ERrs (Expression, Expression)
right residual

EDia (Expression, Expression)
diamond <>

ECps (Expression, Expression)
composition ;

ERad (Expression, Expression)
relative addition !

EPrd (Expression, Expression)
cartesian product *

EKl0 Expression
Rfx.Trn closure * (Kleene star)

EKl1 Expression
Transitive closure + (Kleene plus)

EFlp Expression
conversion (flip, wok) ~

ECpl Expression
Complement

EBrk Expression
bracketed expression ( ... )

EDcD Declaration
simple declaration

EDcI A_Concept
Identity relation

EEps A_Concept Sign
Epsilon relation (introduced by the system to ensure we compare concepts by equality only.

EDcV Sign
Cartesian product relation

EMp1 String A_Concept
constant (string between single quotes)

getExpressionRelation :: Expression -> Maybe (A_Concept, Declaration, A_Concept, Bool)

data A_Concept

Constructors

PlainConcept

Fields

cptnm :: String

PlainConcept nm represents the set of instances cs by name nm.

ONE
The universal Singleton: I[Anything] = V['Anything'*'Anything']

data A_Markup

Constructors

A_Markup

Fields

amLang :: Lang

amFormat :: PandocFormat

amPandoc :: [Block]

data AMeaning

Constructors

AMeaning

Fields

ameaMrk :: [A_Markup]

data A_RoleRule

Constructors

A_RoleRule

Fields

arRoles :: [Role]

arRules :: [String]

arPos :: Origin

data A_RoleRelation

Constructors

RR

Fields

rrRoles :: [Role]

name of a role

rrRels :: [Declaration]

name of a Relation

rrPos :: Origin

position in the Ampersand script

data Sign

Constructors

Sign A_Concept A_Concept

data Population

Constructors

PRelPopu

Fields

popdcl :: Declaration

popps :: Pairs

PCptPopu

Fields

popcpt :: A_Concept

popas :: [String]

class Association rel where

Minimal complete definition

sign

Methods

source, target :: rel -> A_Concept

sign :: rel -> Sign

isEndo :: rel -> Bool

data PAclause

Constructors

CHC

Fields

paCls :: [PAclause]

paMotiv :: [(Expression, [Rule])]

GCH

Fields

paGCls :: [(InsDel, Expression, PAclause)]

paMotiv :: [(Expression, [Rule])]

ALL

Fields

paCls :: [PAclause]

paMotiv :: [(Expression, [Rule])]

Do

Fields

paSrt :: InsDel

paTo :: Declaration

paDelta :: Expression

paMotiv :: [(Expression, [Rule])]

New

Fields

paCpt :: A_Concept

paCl :: String -> PAclause

paMotiv :: [(Expression, [Rule])]

Rmv

Fields

paCpt :: A_Concept

paCl :: String -> PAclause

paMotiv :: [(Expression, [Rule])]

Nop

Fields

paMotiv :: [(Expression, [Rule])]

Blk

Fields

paMotiv :: [(Expression, [Rule])]

Let

Fields

paExpr :: PAclause

paBody :: PAclause -> PAclause

paMotiv :: [(Expression, [Rule])]

Ref

Fields

paVar :: String

data Event

Constructors

On

Fields

eSrt :: InsDel

eDcl :: Declaration

data ECArule

Constructors

ECA

Fields

ecaTriggr :: Event

ecaDelta :: Declaration

ecaAction :: PAclause

ecaNum :: Int

data InsDel

Constructors

Ins

Del

data Conjunct

Constructors

Cjct

Fields

rc_id :: String

rc_orgRules :: [Rule]

rc_conjunct :: Expression

rc_dnfClauses :: [DnfClause]

data DnfClause

Constructors

Dnf

Fields

antcs :: [Expression]

conss :: [Expression]

(<==>) :: Poset a => a -> a -> Bool

Is comparable to.

join :: Sortable a => a -> a -> a

meet :: Sortable a => a -> a -> a

greatest :: (Show a, Sortable a) => [a] -> a

takes the greatest a of comparables

least :: Sortable a => [a] -> a

takes the least a of comparables if there is only one

maxima :: Sortable a => [a] -> [a]

takes all a without anything larger

minima :: Sortable a => [a] -> [a]

takes all a without anything less

sortWith :: (Show b, Poset b) => (b -> [[b]], [b]) -> (a -> b) -> [a] -> [a]

smallerConcepts :: [A_Gen] -> A_Concept -> [A_Concept]

this function takes all generalisation relations from the context and a concept and delivers a list of all concepts that are more specific than the given concept. If there are no cycles in the generalization graph, cpt cannot be an element of smallerConcepts gens cpt.

largerConcepts :: [A_Gen] -> A_Concept -> [A_Concept]

this function takes all generalisation relations from the context and a concept and delivers a list of all concepts that are more generic than the given concept.

rootConcepts :: [A_Gen] -> [A_Concept] -> [A_Concept]

this function returns the most generic concepts in the class of a given concept

genericAndSpecifics :: A_Gen -> [(A_Concept, A_Concept)]

showSign :: Association a => a -> String

aMarkup2String :: A_Markup -> String

module Database.Design.Ampersand.Core.ParseTree

(.==.) :: Expression -> Expression -> Expression infixl 1

(.|-.) :: Expression -> Expression -> Expression infixl 1

(./\.) :: Expression -> Expression -> Expression infixl 2

(.\/.) :: Expression -> Expression -> Expression infixl 2

(.-.) :: Expression -> Expression -> Expression infixl 4

(./.) :: Expression -> Expression -> Expression infixl 6

(.\.) :: Expression -> Expression -> Expression infixl 6

(.<>.) :: Expression -> Expression -> Expression infixl 6

(.:.) :: Expression -> Expression -> Expression infixl 8

(.!.) :: Expression -> Expression -> Expression infixl 8

(.*.) :: Expression -> Expression -> Expression infixl 8