Database.Design.Ampersand.Core.ParseTree

Documentation

data P_Context

Constructors

PCtx

Fields

ctx_nm :: String

The name of this context

ctx_pos :: [Origin]

The origin of the context. A context can be a merge of a file including other files c.q. a list of Origin.

ctx_lang :: Lang

The default language specified on the top-level context

ctx_markup :: Maybe PandocFormat

The default markup format for free text in this context

ctx_thms :: [String]

Names of patterns/processes to be printed in the functional specification. (For partial documents.)

ctx_pats :: [P_Pattern]

The patterns defined in this context

ctx_rs :: [P_Rule TermPrim]

All user defined rules in this context, but outside patterns and outside processes

ctx_ds :: [P_Declaration]

The relations defined in this context, outside the scope of patterns

ctx_cs :: [ConceptDef]

The concept definitions defined in this context, outside the scope of patterns

ctx_ks :: [P_IdentDef]

The identity definitions defined in this context, outside the scope of patterns

ctx_rrules :: [P_RoleRule]

The MAINTAIN definitions defined in this context, outside the scope of patterns

ctx_rrels :: [P_RoleRelation]

The assignment of roles to Relations. (EDITS statements)

ctx_vs :: [P_ViewDef]

The view definitions defined in this context, outside the scope of patterns

ctx_gs :: [P_Gen]

The gen definitions defined in this context, outside the scope of patterns

ctx_ifcs :: [P_Interface]

The interfaces defined in this context

ctx_ps :: [PPurpose]

The purposes defined in this context, outside the scope of patterns and processes

ctx_pops :: [P_Population]

The populations defined in this context

ctx_sql :: [P_ObjectDef]

user defined sqlplugs, taken from the Ampersand script

ctx_php :: [P_ObjectDef]

user defined phpplugs, taken from the Ampersand script

ctx_metas :: [Meta]

generic meta information (name/value pairs) that can be used for experimenting without having to modify the adl syntax

mergeContexts :: P_Context -> P_Context -> P_Context

data Meta

Constructors

Meta

Fields

mtPos :: Origin

mtObj :: MetaObj

mtName :: String

mtVal :: String

data MetaObj

Constructors

ContextMeta

data P_RoleRelation

A RoleRelation rs means that any role in 'rrRoles rs' may edit any Relation in 'rrInterfaces rs'

Constructors

P_RR

Fields

rr_Pos :: Origin

position in the Ampersand script

rr_Roles :: [Role]

list of roles

rr_Rels :: [P_NamedRel]

list of named relations

data P_RoleRule

Constructors

Maintain

Fields

mPos :: Origin

position in the Ampersand script

mRoles :: [Role]

name of a role

mRules :: [String]

name of a Rule

data Role

Constructors

Role String

data P_Pattern

Constructors

P_Pat

Fields

pt_pos :: Origin

the starting position in the file in which this pattern was declared.

pt_nm :: String

Name of this pattern

pt_rls :: [P_Rule TermPrim]

The user defined rules in this pattern

pt_gns :: [P_Gen]

The generalizations defined in this pattern

pt_dcs :: [P_Declaration]

The relations that are declared in this pattern

pt_RRuls :: [P_RoleRule]

The assignment of roles to rules.

pt_RRels :: [P_RoleRelation]

The assignment of roles to Relations.

pt_cds :: [ConceptDef]

The concept definitions defined in this pattern

pt_ids :: [P_IdentDef]

The identity definitions defined in this pattern

pt_vds :: [P_ViewDef]

The view definitions defined in this pattern

pt_xps :: [PPurpose]

The purposes of elements defined in this pattern

pt_pop :: [P_Population]

The populations that are local to this pattern

pt_end :: Origin

the end position in the file in which this pattern was declared.

data P_Declaration

Constructors

P_Sgn

Fields

dec_nm :: String

the name of the declaration

dec_sign :: P_Sign

the type. Parser must guarantee it is not empty.

dec_prps :: Props

the user defined multiplicity properties (Uni, Tot, Sur, Inj) and algebraic properties (Sym, Asy, Trn, Rfx)

dec_pragma :: [String]

Three strings, which form the pragma. E.g. if pragma consists of the three strings: "Person ", " is married to person ", and " in Vegas." ^ then a tuple (Peter,Jane) in the list of links means that Person Peter is married to person Jane in Vegas.

dec_Mean :: [PMeaning]

the optional meaning of a declaration, possibly more than one for different languages.

dec_popu :: Pairs

the list of tuples, of which the relation consists.

dec_fpos :: Origin

the position in the Ampersand source file where this declaration is declared. Not all decalartions come from the ampersand souce file.

dec_plug :: Bool

if true, this relation may not be stored in or retrieved from the standard database (it should be gotten from a Plug of some sort instead)

data Term a

Constructors

Prim a

PEqu Origin (Term a) (Term a)
equivalence =

PImp Origin (Term a) (Term a)
implication |-

PIsc Origin (Term a) (Term a)
intersection /

PUni Origin (Term a) (Term a)
union /

PDif Origin (Term a) (Term a)
difference -

PLrs Origin (Term a) (Term a)
left residual /

PRrs Origin (Term a) (Term a)
right residual

PDia Origin (Term a) (Term a)
diamond <>

PCps Origin (Term a) (Term a)
composition ;

PRad Origin (Term a) (Term a)
relative addition !

PPrd Origin (Term a) (Term a)
cartesian product *

PKl0 Origin (Term a)
Rfx.Trn closure * (Kleene star)

PKl1 Origin (Term a)
Transitive closure + (Kleene plus)

PFlp Origin (Term a)
conversion (flip, wok) ~

PCpl Origin (Term a)
Complement

PBrk Origin (Term a)
bracketed expression ( ... )

data TermPrim

Constructors

PI Origin
identity element without a type At parse time, there may be zero or one element in the list of concepts. Reason: when making eqClasses, the least element of that class is used as a witness of that class to know whether an eqClass represents a concept, we only look at its witness By making Pid the first in the data decleration, it becomes the least element for "deriving Ord".

Pid Origin P_Concept
identity element restricted to a type

Patm Origin String (Maybe P_Concept)
an atom, possibly with a type

PVee Origin
the complete relation, of which the type is yet to be derived by the type checker.

Pfull Origin P_Concept P_Concept
the complete relation, restricted to a type. At parse time, there may be zero, one or two elements in the list of concepts.

PNamedR P_NamedRel

data P_NamedRel

Constructors

PNamedRel Origin String (Maybe P_Sign)

data PairView a

Constructors

PairView

Fields

ppv_segs :: [PairViewSegment a]

data PairViewSegment a

Constructors

PairViewText

Fields

pvsOrg :: Origin

pvsStr :: String

PairViewExp

Fields

pvsOrg :: Origin

pvsSoT :: SrcOrTgt

pvsExp :: a

newtype PairViewTerm a

the newtype to make it possible for a PairView to be disambiguatable: it must be of the form "d a" instead of "d (Term a)"

Constructors

PairViewTerm (PairView (Term a))

newtype PairViewSegmentTerm a

Constructors

PairViewSegmentTerm (PairViewSegment (Term a))

data SrcOrTgt

Constructors

Src

Tgt

isSrc :: SrcOrTgt -> Bool

data P_Rule a

Constructors

P_Ru

Fields

rr_fps :: Origin

Position in the Ampersand file

rr_nm :: String

Name of this rule

rr_exp :: Term a

The rule expression

rr_mean :: [PMeaning]

User-specified meanings, possibly more than one, for multiple languages.

rr_msg :: [PMessage]

User-specified violation messages, possibly more than one, for multiple languages.

rr_viol :: Maybe (PairView (Term a))

Custom presentation for violations, currently only in a single language

data ConceptDef

Constructors

Cd

Fields

cdpos :: Origin

The position of this definition in the text of the Ampersand source (filename, line number and column number).

cdcpt :: String

The name of the concept for which this is the definition. If there is no such concept, the conceptdefinition is ignored.

cdplug :: Bool

Whether the user specifically told Ampersand not to store this concept in the database

cddef :: String

The textual definition of this concept.

cdtyp :: String

The (SQL) type of this concept.

cdref :: String

A label meant to identify the source of the definition. (useful as LaTeX' symbolic reference)

cdfrom :: String

The name of the pattern or context in which this concept definition was made

data P_Population

Constructors

P_RelPopu

Fields

p_orig :: Origin

p_rnme :: String

p_popps :: Pairs

P_TRelPop

Fields

p_orig :: Origin

p_rnme :: String

p_type :: P_Sign

p_popps :: Pairs

P_CptPopu

Fields

p_orig :: Origin

p_cnme :: String

p_popas :: [String]

type P_ObjectDef = P_ObjDef TermPrim

type P_SubInterface = P_SubIfc TermPrim

data P_Interface

Constructors

P_Ifc

Fields

ifc_Name :: String

the name of the interface

ifc_Class :: Maybe String

the class of the interface

ifc_Params :: [P_NamedRel]

a list of named relations that are editable within this interface.

ifc_Args :: [[String]]

a list of arguments for code generation.

ifc_Roles :: [Role]

a list of roles that may use this interface

ifc_Obj :: P_ObjectDef

the context expression (mostly: I[c])

ifc_Pos :: Origin

ifc_Prp :: String

data P_IClass

Constructors

P_IClass

Fields

iclass_name :: String

data P_ObjDef a

Constructors

P_Obj

Fields

obj_nm :: String

view name of the object definition. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface if it is not an empty string.

obj_pos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number)

obj_ctx :: Term a

this expression describes the instances of this object, related to their context.

obj_mView :: Maybe String

The view that should be used for this object

obj_msub :: Maybe (P_SubIfc a)

the attributes, which are object definitions themselves.

obj_strs :: [[String]]

directives that specify the interface.

data P_SubIfc a

Constructors

P_Box

Fields

si_ori :: Origin

si_class :: Maybe String

si_box :: [P_ObjDef a]

P_InterfaceRef

Fields

si_ori :: Origin

si_str :: String

type P_IdentDef = P_IdentDf TermPrim

data P_IdentDf a

Constructors

P_Id

Fields

ix_pos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

ix_lbl :: String

the name (or label) of this Identity. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

ix_cpt :: P_Concept

this expression describes the instances of this object, related to their context

ix_ats :: [P_IdentSegmnt a]

the constituent segments of this identity. TODO: refactor to a list of terms

type P_IdentSegment = P_IdentSegmnt TermPrim

data P_IdentSegmnt a

Constructors

P_IdentExp

Fields

ks_obj :: P_ObjDef a

type P_ViewDef = P_ViewD TermPrim

type P_ViewSegment = P_ViewSegmt TermPrim

data ViewHtmlTemplate

Constructors

ViewHtmlTemplateFile String

data P_ViewD a

Constructors

P_Vd

Fields

vd_pos :: Origin

position of this definition in the text of the Ampersand source file (filename, line number and column number).

vd_lbl :: String

the name (or label) of this View. The label has no meaning in the Compliant Service Layer, but is used in the generated user interface. It is not an empty string.

vd_cpt :: P_Concept

the concept for which this view is applicable

vd_isDefault :: Bool

whether or not this is the default view for the concept

vd_html :: Maybe ViewHtmlTemplate

the html template for this view (not required since we may have other kinds of views as well in the future) , vd_text :: Maybe P_ViewText -- Future extension

vd_ats :: [P_ViewSegmt a]

the constituent segments of this view.

data P_ViewSegmt a

Constructors

P_ViewExp

Fields

vs_obj :: P_ObjDef a

P_ViewText

Fields

vs_txt :: String

P_ViewHtml

Fields

vs_htm :: String

data PPurpose

Constructors

PRef2

Fields

pexPos :: Origin

pexObj :: PRef2Obj

pexMarkup :: P_Markup

pexRefIDs :: [String]

data PRef2Obj

Constructors

PRef2ConceptDef String

PRef2Declaration P_NamedRel

PRef2Rule String

PRef2IdentityDef String

PRef2ViewDef String

PRef2Pattern String

PRef2Interface String

PRef2Context String

PRef2Fspc String

newtype PMeaning

Constructors

PMeaning P_Markup

newtype PMessage

Constructors

PMessage P_Markup

data P_Concept

Constructors

PCpt
The name of this Concept

Fields

p_cptnm :: String

P_Singleton

data P_Sign

Constructors

P_Sign

Fields

pSrc :: P_Concept

pTgt :: P_Concept

data P_Gen

Constructors

P_Cy

Fields

gen_fp :: Origin

Position in the Ampersand file

gen_spc :: P_Concept

Left hand side concept expression

gen_rhs :: [P_Concept]

Right hand side concept expression

PGen

Fields

gen_fp :: Origin

Position in the Ampersand file

gen_spc :: P_Concept

Left hand side concept expression

gen_gen :: P_Concept

generic concept

data Lang

Constructors

Dutch

English

data P_Markup

Constructors

P_Markup

Fields

mLang :: Maybe Lang

mFormat :: Maybe PandocFormat

mString :: String

data PandocFormat

Constructors

HTML

ReST

LaTeX

Markdown

data Prop

Constructors

Uni
univalent

Inj
injective

Sur
surjective

Tot
total

Sym
symmetric

Asy
antisymmetric

Trn
transitive

Rfx
reflexive

Irf
irreflexive

Aut
automatically computed (NOTE: this is a hacky way to denote these until we have appropriate syntax)

Prop
PROP keyword, later replaced by [Sym, Asy]

type Props = [Prop]

normalizeProps :: [Prop] -> [Prop]

module Database.Design.Ampersand.Input.ADL1.FilePos

module Database.Design.Ampersand.ADL1.Pair

gen_concs :: P_Gen -> [P_Concept]