Database.Design.Ampersand.Core.Poset.Internal

Documentation

fatal :: Int -> String -> a

data Ordering

Constructors

LT

EQ

GT

CP

NC

totalOrder :: Ordering -> Ordering

Internal-use function to convert our Ordering to the ordinary one.

partialOrder :: Ordering -> Ordering

Internal-use function to convert the ordinary Ordering to ours.

class Eq a => Poset a where

Class for partially ordered data types. Instances should satisfy the following laws for all values a, b and c:

a <= a.

a <= b and b <= a implies a == b.

a <= b and b <= c implies a <= c.

But note that the floating point instances don't satisfy the first rule.

Minimal complete definition: compare or <=.

Minimal complete definition

Nothing

Methods

compare :: a -> a -> Ordering

(<==>) :: a -> a -> Bool

Is comparable to.

(</=>) :: a -> a -> Bool

Is not comparable to.

(<) :: a -> a -> Bool

(<=) :: a -> a -> Bool

(>=) :: a -> a -> Bool

(>) :: a -> a -> Bool

class Poset a => Sortable a where

Class for partially ordered data types where sorting makes sense. This includes all totally ordered sets and floating point types. Instances should satisfy the following laws:

The set of elements for which isOrdered returns true is totally ordered.

The max (or min) of an insignificant element and a significant element is the significant one.

The result of sorting a list should contain only significant elements.

max a b = max b a

min a b = min b a

The idea comes from floating point types, where non-comparable elements (NaNs) are the exception rather than the rule. For these types, we can define max, min and sortBy to ignore insignificant elements. Thus, a sort of floating point values will discard all NaNs and order the remaining elements.

Minimal complete definition: isOrdered

Methods

sortBy :: (a -> a -> Ordering) -> [a] -> [a]

join :: a -> a -> a

meet :: a -> a -> a

class Poset a => Ord a

Class for totally ordered data types. Instances should satisfy isOrdered a = True for all a.