Database.Design.Ampersand.Core.ToMeta

Documentation

toMeta :: MakeMeta a => Options -> a -> a

When dealing with meta-stuff for Ampersand, (Like makeGenerics, makeRAP), the names of Concepts should be different than normal, user-defined Concepts. This function modifies everything in the context to reflect that.