Database.Design.Ampersand.FSpec.FPA

Documentation

data FPA

Constructors

FPA

Fields

dataModelFPA :: ([FP], Int)

userTransactionFPA :: ([FP], Int)

data FP

Constructors

FP

Fields

fpType :: FPType

fpName :: String

fpComplexity :: Complexity

data FPType

Constructors

ILGV
bevat permanente, voor de gebruiker relevante gegevens. De gegevens worden door het systeem gebruikt en onderhouden. Onder "onderhouden" verstaat FPA het toevoegen, wijzigen of verwijderen van gegevens.

KGV
bevat permanente, voor de gebruiker relevante gegevens. Deze gegevens worden door het systeem gebruikt, maar worden door een ander systeem onderhouden (voor dat andere systeem is het dus een ILGV).

IF
verwerkt gegevens in een ILGV van het systeem. (dus create, update en delete functies)

UF
presenteert gegevens uit het systeem. Voorbeelden: het afdrukken van alle debiteuren; het aanmaken van facturen; het aanmaken van een diskette met betalingsopdrachten; het medium is hierbij niet van belang: papier, scherm, magneetband, datacom, enzovoorts.

OF
is een speciaal (eenvoudig) soort uitvoerfunctie. Een opvraagfunctie presenteert gegevens uit het systeem op basis van een uniek identificerend zoekgegeven, waarbij geen aanvullende bewerkingen (zoals berekeningen of het bijwerken van een gegevensverzameling) plaats hebben. Voorbeeld: Het tonen van de gegevens van de klant met klantnummer 123456789.

class ShowLang a where

Methods

showLang :: Lang -> a -> String

fpAnalyze :: FSpec -> FPA

fpVal :: FP -> Int

Valuing of function points according to par. 3.9 (UK) or par. 2.9 (NL), see http:/www.nesma.nlsectiefpahoefpa.asp

fpaPlugInfo :: PlugInfo -> Maybe FP

fpaInterface :: Interface -> FP