Database.Design.Ampersand.FSpec.FSpec

Description

The intentions behind FSpec (SJ 30 dec 2008): Generation of functional specifications is the core functionality of Ampersand. All items in a specification are generated into the following data structure, FSpec. It is built by compiling an Ampersand script and translating that to FSpec. In the future, other ways of filling FSpec are foreseen. All generators (such as the code generator, the proof generator, the atlas generator, etc.) are merely different ways to show FSpec.

Documentation

data FSpec

Constructors

FSpec

Fields

fsName :: String

The name of the specification, taken from the Ampersand script

originalContext :: A_Context

the original context. (for showADL)

getOpts :: Options

The command line options that were used when this FSpec was compiled by Ampersand.

fspos :: [Origin]

The origin of the FSpec. An FSpec can be a merge of a file including other files c.q. a list of Origin.

themes :: [String]

The names of patterns/processes to be printed in the functional specification. (for making partial documentation)

pattsInScope :: [Pattern]

rulesInScope :: [Rule]

declsInScope :: [Declaration]

concsInScope :: [A_Concept]

cDefsInScope :: [ConceptDef]

gensInScope :: [A_Gen]

fsLang :: Lang

The default language for this specification (always specified, so no Maybe here!).

vplugInfos :: [PlugInfo]

All plugs defined in the Ampersand script

plugInfos :: [PlugInfo]

All plugs (defined and derived)

interfaceS :: [Interface]

All interfaces defined in the Ampersand script

interfaceG :: [Interface]

All interfaces derived from the basic ontology (the Lonneker interface)

fSwitchboard :: Fswitchboard

The code to be executed to maintain the truth of invariants

fDeriveProofs :: Blocks

The proofs in Pandoc format

fActivities :: [Activity]

generated: One Activity for every ObjectDef in interfaceG and interfaceS

fRoleRels :: [(Role, Declaration)]

the relation saying which roles may change the population of which relation.

fRoleRuls :: [(Role, Rule)]

the relation saying which roles may change the population of which relation.

fRoles :: [Role]

All roles mentioned in this context.

fallRules :: [Rule]

vrules :: [Rule]

All user defined rules that apply in the entire FSpec

grules :: [Rule]

All rules that are generated: multiplicity rules and identity rules

invariants :: [Rule]

All invariant rules

allUsedDecls :: [Declaration]

All relations that are used in the fSpec

allDecls :: [Declaration]

All relations that are declared in the fSpec

vrels :: [Declaration]

All user defined and generated relations plus all defined and computed totals. The generated relations are all generalizations and one declaration for each signal.

allConcepts :: [A_Concept]

All concepts in the fSpec

kernels :: [[A_Concept]]

All concepts, grouped by their classifications

vIndices :: [IdentityDef]

All keys that apply in the entire FSpec

vviews :: [ViewDef]

All views that apply in the entire FSpec

vgens :: [A_Gen]

All gens that apply in the entire FSpec

vconjs :: [Conjunct]

All conjuncts generated (by ADL2FSpec)

allConjsPerRule :: [(Rule, [Conjunct])]

Maps each rule onto the conjuncts it consists of (note that a single conjunct may be part of several rules)

allConjsPerDecl :: [(Declaration, [Conjunct])]

Maps each declaration to the conjuncts it appears in

allConjsPerConcept :: [(A_Concept, [Conjunct])]

Maps each concept to the conjuncts it appears in (as source or targent of a constituent relation)

vquads :: [Quad]

All quads generated (by ADL2FSpec)

vEcas :: [ECArule]

All ECA rules generated (by ADL2FSpec)

fsisa :: [(A_Concept, A_Concept)]

generated: The data structure containing the generalization structure of concepts

vpatterns :: [Pattern]

All patterns taken from the Ampersand script

conceptDefs :: [ConceptDef]

All concept definitions defined throughout a context, including those inside patterns and processes

fSexpls :: [Purpose]

All purposes that have been declared at the top level of the current specification, but not in the processes, patterns and interfaces.

metas :: [Meta]

All meta relations from the entire context

crudInfo :: CrudInfo

Information for CRUD matrices

initialPops :: [Population]

All user defined populations of relations and concepts

allAtoms :: [Atom]

allLinks :: [A_Pair]

initialConjunctSignals :: [(Conjunct, [Paire])]

All conjuncts that have process-rule violations.

allViolations :: [(Rule, [Paire])]

All invariant rules with violations.

allExprs :: [Expression]

All expressions in the fSpec

allSigns :: [Sign]

All Signs in the fSpec

concDefs :: FSpec -> A_Concept -> [ConceptDef]

data Atom

Constructors

Atom

Fields

atmRoots :: [A_Concept]

atmIn :: [A_Concept]

atmVal :: String

data A_Pair

Constructors

Pair

Fields

lnkDcl :: Declaration

lnkLeft :: Atom

lnkRight :: Atom

data Fswitchboard

A list of ECA rules, which is used for automated functionality.

Constructors

Fswtch

Fields

fsbEvIn :: [Event]

fsbEvOut :: [Event]

fsbConjs :: [(Rule, Expression)]

fsbECAs :: [ECArule]

data Quad

Constructors

Quad

Fields

qDcl :: Declaration

qRule :: Rule

qConjuncts :: [Conjunct]

data FSid

Constructors

FS_id String

data ECArule

Constructors

ECA

Fields

ecaTriggr :: Event

ecaDelta :: Declaration

ecaAction :: PAclause

ecaNum :: Int

data Activity

Constructors

Act

Fields

actRule :: Rule

actTrig :: [Declaration]

actAffect :: [Declaration]

actQuads :: [Quad]

actEcas :: [ECArule]

actPurp :: [Purpose]

data PlugSQL

Constructors

TblSQL
stores a related collection of relations: a kernel of concepts and attribute relations of this kernel i.e. a list of SqlField given some A -> [target r | r::A*B,isUni r,isTot r, isInj r] ++ [target r | r::A*B,isUni r, not(isTot r), not(isSur r)] kernel = A closure of concepts A,B for which there exists a r::A->B[INJ] (r=fldexpr of kernel field holding instances of B, in practice r is I or a makeRelation(flipped declaration)) attribute relations = All concepts B, A in kernel for which there exists a r::A*B[UNI] and r not TOT and SUR (r=fldexpr of attMor field, in practice r is a makeRelation(declaration))

Fields

sqlname :: String

fields :: [SqlField]

the first field is the concept table of the most general concept (e.g. Person) then follow concept tables of specializations. Together with the first field this is called the "kernel" the remaining fields represent attributes.

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkpTbl :: [(Expression, SqlField, SqlField)]

lookup table that links concepts to column names in the plug (kernel+attRels) mLkpTbl is een lijst met relaties die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

BinSQL
stores one relation r in two ordered columns i.e. a tuple of SqlField -> (source r,target r) with (fldexpr=I/r;r~, fldexpr=r) (note: if r TOT then (I/r;r~ = I). Thus, the concept (source r) is stored in this plug too) with tblcontents = [[Just x,Just y] |(x,y)<-contents r]. Typical for BinSQL is that it has exactly two columns that are not unique and may not contain NULL values

Fields

sqlname :: String

columns :: (SqlField, SqlField)

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkp :: Expression

ScalarSQL
stores one concept c in one column i.e. a SqlField -> c with tblcontents = [[Just x] |(x,_)<-contents c]. Typical for ScalarSQL is that it has exactly one column that is unique and may not contain NULL values i.e. fldexpr=I[c]

Fields

sqlname :: String

sqlColumn :: SqlField

cLkp :: A_Concept

plugFields :: PlugSQL -> [SqlField]

lookupCpt :: FSpec -> A_Concept -> [(PlugSQL, SqlField)]

This returns all columntable pairs that serve as a concept table for cpt. When addingremoving atoms, all of these columns need to be updated

getConceptTableFor :: FSpec -> A_Concept -> String

metaValues :: String -> FSpec -> [String]

data SqlField

Constructors

Fld

Fields

fldname :: String

fldexpr :: Expression

De target van de expressie geeft de waarden weer in de SQL-tabel-kolom.

fldtype :: SqlType

flduse :: SqlFieldUsage

fldnull :: Bool

True if there can be empty field-values (intended for data dictionary of DB-implementation)

flduniq :: Bool

True if all field-values are unique? (intended for data dictionary of DB-implementation)

class Object a where

Methods

concept :: a -> A_Concept

attributes :: a -> [ObjectDef]

contextOf :: a -> Expression

data PlugInfo

Constructors

InternalPlug PlugSQL

ExternalPlug ObjectDef

data SqlType

Constructors

SQLChar Int

SQLBlob

SQLPass

SQLSingle

SQLDouble

SQLText

SQLuInt Int

SQLsInt Int

SQLId

SQLVarchar Int

SQLBool

data SqlFieldUsage

Constructors

TableKey Bool A_Concept

ForeignKey A_Concept

PlainAttr

getGeneralizations :: FSpec -> A_Concept -> [A_Concept]

getSpecializations :: FSpec -> A_Concept -> [A_Concept]

lookupView :: FSpec -> String -> ViewDef

getDefaultViewForConcept :: FSpec -> A_Concept -> Maybe ViewDef

data Conjunct

Constructors

Cjct

Fields

rc_id :: String

rc_orgRules :: [Rule]

rc_conjunct :: Expression

rc_dnfClauses :: [DnfClause]

data DnfClause

Constructors

Dnf

Fields

antcs :: [Expression]

conss :: [Expression]

dnf2expr :: DnfClause -> Expression

notCpl :: Expression -> Expression

class Language a where

Minimal complete definition

relsDefdIn, udefrules, identities, viewDefs, gens, patterns

Methods

relsDefdIn

Arguments

:: a

-> [Declaration]
all relations that are declared in the scope of this viewpoint. These are user defined relations and all generated relarations, i.e. one relation for each GEN and one for each signal rule. Don't confuse relsDefdIn with relsUsedIn, which gives the relations that are used in a.)

udefrules

Arguments

:: a

-> [Rule]
all user defined rules that are maintained within this viewpoint, which are not multiplicity- and not identity rules.

multrules

Arguments

:: a

-> [Rule]
all multiplicityrules that are maintained within this viewpoint.

identityRules :: a -> [Rule]

allRules :: a -> [Rule]

identities

Arguments

:: a

-> [IdentityDef]
all keys that are defined in a

viewDefs

Arguments

:: a

-> [ViewDef]
all views that are defined in a

gens

Arguments

:: a

-> [A_Gen]
all generalizations that are valid within this viewpoint

patterns

Arguments

:: a

-> [Pattern]
all patterns that are used in this viewpoint