Database.Design.Ampersand.FSpec.Graphic.ClassDiagram

Documentation

data ClassDiag

Constructors

OOclassdiagram

Fields

cdName :: String

classes :: [Class]

assocs :: [Association]

aggrs :: [Aggregation]

geners :: [Generalization]

ooCpts :: [A_Concept]

data Class

Constructors

OOClass

Fields

clName :: String

name of the class

clcpt :: Maybe (A_Concept, [Purpose])

Main concept of the class. (link tables do not have a main concept)

clAtts :: [CdAttribute]

Attributes of the class

clMths :: [Method]

Methods of the class

data CdAttribute

Constructors

OOAttr

Fields

attNm :: String

name of the attribute

attTyp :: String

type of the attribute (Concept name or built-in type)

attOptional :: Bool

says whether the attribute is optional

data Association

Constructors

OOAssoc

Fields

assSrc :: String

source: the name of the source class

assSrcPort :: String

the name of the attribute in the source class

asslhm :: Multiplicities

left hand side multiplicities

asslhr :: String

left hand side role

assTgt :: String

target: the name of the target class

assrhm :: Multiplicities

right hand side multiplicities

assrhr :: String

right hand side role

asspurp :: [Purpose]

purposes of this association

assmean :: Maybe A_Markup

meaning, if available.

data Aggregation

Constructors

OOAggr

Fields

aggDel :: Deleting

aggChild :: A_Concept

aggParent :: A_Concept

data Generalization

Constructors

OOGener

Fields

genAgen :: A_Gen

data Deleting

Constructors

Open

Close

data Method

Constructors

OOMethodC String [CdAttribute]

OOMethodR String [CdAttribute]

OOMethodS String [CdAttribute]

OOMethodU String [CdAttribute]

OOMethodD String

OOMethod String [CdAttribute] String

data Multiplicities

Constructors

Mult MinValue MaxValue

data MinValue

Constructors

MinZero

MinOne

data MaxValue

Constructors

MaxOne

MaxMany

clAnalysis :: FSpec -> ClassDiag

This function makes the classification diagram. It focuses on generalizations and specializations.

cdAnalysis :: FSpec -> ClassDiag

This function, cdAnalysis, generates a conceptual data model. It creates a class diagram in which generalizations and specializations remain distinct entity types. This yields more classes than plugs2classdiagram does, as plugs contain their specialized concepts. Properties and identities are not shown.

tdAnalysis :: FSpec -> ClassDiag

This function generates a technical data model. It is based on the plugs that are calculated.

classdiagram2dot :: Options -> ClassDiag -> DotGraph String

translate a ClassDiagram to a DotGraph, so it can be used to show it as a picture.