Database.Design.Ampersand.FSpec.Graphic.Graphics

Documentation

makePicture :: FSpec -> PictureReq -> Picture

writePicture :: Options -> Picture -> IO ()

data Picture

Constructors

Pict

Fields

pType :: PictureReq

the type of the picture

scale :: String

a scale factor, intended to pass on to LaTeX, because Pandoc seems to have a problem with scaling.

dotSource :: DotGraph String

the string representing the .dot

dotProgName :: GraphvizCommand

the name of the program to use ("dot" or "neato" or "fdp" or Sfdp)

caption :: String

a human readable name of this picture

data PictureReq

Constructors

PTClassDiagram

PTRelsUsedInPat Pattern

PTDeclaredInPat Pattern

PTConcept A_Concept

PTSwitchBoard Activity

PTFinterface Activity

PTIsaInPattern Pattern

PTSingleRule Rule

PTLogicalDM

PTTechnicalDM

imagePath

Arguments

:: ReferableFromPandoc a

=> Options

-> a

-> FilePath
the full file path to the image file