Database.Design.Ampersand.FSpec.Motivations

Documentation

class Named a => Motivated a where

Minimal complete definition

explForObj, explanations

Methods

purposeOf

Arguments

:: FSpec

-> Lang

-> a

-> Maybe [Purpose]
explains the purpose of a, i.e. the reason why a exists. The purpose could be either given by the user, or generated by Ampersand. Multiple purposes are allowed for the following reasons: * Different purposes from different sources make me want to document them all. * Combining two overlapping scripts from (i.e. from different authors) may cause multiple purposes.

explForObj

Arguments

:: a

-> ExplObj

-> Bool
Given an Explainable object and an ExplObj, return TRUE if they concern the identical object.

explanations

Arguments

:: a

-> [Purpose]
The explanations that are defined inside a (including that of a itself)

purposesDefinedIn

Arguments

:: FSpec

-> Lang

-> a

-> [Purpose]
The explanations that are defined inside a (including that of a itself)

class Meaning a where

Minimal complete definition

meaning

Methods

meaning :: Lang -> a -> Maybe A_Markup

meaning2Blocks :: Lang -> a -> [Block]