Database.Design.Ampersand.FSpec.Plug

Documentation

class (Named p, Eq p, Show p) => Plugable p where

Methods

makePlug :: PlugInfo -> p

data PlugInfo

Constructors

InternalPlug PlugSQL

ExternalPlug ObjectDef

data SqlField

Constructors

Fld

Fields

fldname :: String

fldexpr :: Expression

De target van de expressie geeft de waarden weer in de SQL-tabel-kolom.

fldtype :: SqlType

flduse :: SqlFieldUsage

fldnull :: Bool

True if there can be empty field-values (intended for data dictionary of DB-implementation)

flduniq :: Bool

True if all field-values are unique? (intended for data dictionary of DB-implementation)

data SqlFieldUsage

Constructors

TableKey Bool A_Concept

ForeignKey A_Concept

PlainAttr

data SqlType

Constructors

SQLChar Int

SQLBlob

SQLPass

SQLSingle

SQLDouble

SQLText

SQLuInt Int

SQLsInt Int

SQLId

SQLVarchar Int

SQLBool

showSQL :: SqlType -> String

plugpath :: PlugSQL -> SqlField -> SqlField -> Maybe Expression

tblcontents :: [A_Gen] -> [Population] -> PlugSQL -> [TblRecord]

fldauto :: SqlField -> Bool

data PlugSQL

Constructors

TblSQL
stores a related collection of relations: a kernel of concepts and attribute relations of this kernel i.e. a list of SqlField given some A -> [target r | r::A*B,isUni r,isTot r, isInj r] ++ [target r | r::A*B,isUni r, not(isTot r), not(isSur r)] kernel = A closure of concepts A,B for which there exists a r::A->B[INJ] (r=fldexpr of kernel field holding instances of B, in practice r is I or a makeRelation(flipped declaration)) attribute relations = All concepts B, A in kernel for which there exists a r::A*B[UNI] and r not TOT and SUR (r=fldexpr of attMor field, in practice r is a makeRelation(declaration))

Fields

sqlname :: String

fields :: [SqlField]

the first field is the concept table of the most general concept (e.g. Person) then follow concept tables of specializations. Together with the first field this is called the "kernel" the remaining fields represent attributes.

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkpTbl :: [(Expression, SqlField, SqlField)]

lookup table that links concepts to column names in the plug (kernel+attRels) mLkpTbl is een lijst met relaties die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

BinSQL
stores one relation r in two ordered columns i.e. a tuple of SqlField -> (source r,target r) with (fldexpr=I/r;r~, fldexpr=r) (note: if r TOT then (I/r;r~ = I). Thus, the concept (source r) is stored in this plug too) with tblcontents = [[Just x,Just y] |(x,y)<-contents r]. Typical for BinSQL is that it has exactly two columns that are not unique and may not contain NULL values

Fields

sqlname :: String

columns :: (SqlField, SqlField)

cLkpTbl :: [(A_Concept, SqlField)]

lookup table that links all kernel concepts to fields in the plug cLkpTbl is een lijst concepten die in deze plug opgeslagen zitten, en hoe je ze eruit kunt halen

mLkp :: Expression

ScalarSQL
stores one concept c in one column i.e. a SqlField -> c with tblcontents = [[Just x] |(x,_)<-contents c]. Typical for ScalarSQL is that it has exactly one column that is unique and may not contain NULL values i.e. fldexpr=I[c]

Fields

sqlname :: String

sqlColumn :: SqlField

cLkp :: A_Concept