Database.Design.Ampersand.FSpec.ToFSpec.Calc

Documentation

deriveProofs :: Options -> A_Context -> Blocks

showProof :: (expr -> Blocks) -> Proof expr -> Blocks

showPrf :: (expr -> String) -> Proof expr -> [String]

assembleECAs :: Options -> A_Context -> [Declaration] -> ([ECArule], [Blocks])

assembleECAs assembles larger chunks of code, because it combines acts that are triggered by the same event.

conjuncts :: Options -> Rule -> [Expression]

genPAclause

Arguments

:: (Declaration -> Bool)
True if a relation may be changed (i.e. is editable)

-> InsDel
the type of action: Insert or Delete

-> Expression
the expression in which a delete or insert takes place

-> Expression
the delta to be inserted or deleted

-> [(Expression, [Rule])]
the motivation, consisting of the conjuncts (traced back to their rules) that are being restored by this code fragment.

-> PAclause

de functie genPAclause beschrijft de voornaamste mogelijkheden om een expressie delta' te verwerken in expr (met tOp'==Ins of tOp==Del) TODO: Vind een wetenschappelijk artikel waar de hier beschreven transformatie uitputtend wordt behandeld. TODO: Deze code is onvolledig en misschien zelfs fout....

commaEngPandoc :: Inline -> [Inline] -> [Inline]

commaNLPandoc :: Inline -> [Inline] -> [Inline]

commaEngPandoc' :: Inlines -> [Inlines] -> Inlines

commaNLPandoc' :: Inlines -> [Inlines] -> Inlines

commaPandocAnd :: Lang -> [Inlines] -> Inlines

quadsOfRules :: Options -> [Rule] -> [Quad]