Database.Design.Ampersand.FSpec.ToFSpec.NormalForms

Documentation

delta :: Sign -> Expression

This delta is meant to be used as a placeholder for inserting or removing links from expressions.

conjNF :: Options -> Expression -> Expression

disjNF :: Options -> Expression -> Expression

normPA :: Options -> PAclause -> PAclause

cfProof :: Options -> Expression -> Proof Expression

dfProof :: Options -> Expression -> Proof Expression

proofPA :: Options -> PAclause -> Proof PAclause

simplify :: Expression -> Expression

cfProofs :: Expression -> [(Expression, Proof Expression)]

dfProofs :: Expression -> [(Expression, Proof Expression)]

makeAllConjs :: Options -> [Rule] -> [Conjunct]

conjuncts :: Options -> Rule -> [Expression]