Database.Design.Ampersand.Input.ADL1.Lexer

Documentation

keywords

Arguments

:: [String]
The keywords

Retrieves a list of keywords accepted by the ampersand language

operators

Arguments

:: [String]
The operators

Retrieves a list of operators accepted by the ampersand language

symbols

Arguments

:: String
The list of symbol characters / [Char]

Retrieves the list of symbols accepted by the ampersand language

lexer

Arguments

:: [Options]
The command line options

-> FilePath
The file name, used for error messages

-> String
The content of the file

-> Either LexerError ([Token], [LexerWarning])
Either an error or a list of tokens and warnings

Runs the lexer

data LexerError

Defines a lexer error

Constructors

LexerError FilePos LexerErrorInfo
The lexer file position and error information

data LexerErrorInfo

Defines the different lexer error types

Constructors

UnterminatedComment
The comment was not terminated

UnterminatedPurpose
The purpose statement was not terminated

UnterminatedAtom
The atom was not terminated

UnexpectedChar Char
The character was unexpected

NonTerminatedString String
The string was not terminated

TooManyClose Char
In UnexpectedClose, first char is the closing bracket we see, second char is the closing bracket we would like to see first e.g. [(1,3] => UnexpectedClose ']' ... ')'

UnexpectedClose Char FilePos Char
An unexpected closing bracket was found

StillOpenAtEOF [(FilePos, Char)]
Brackets were not closed at the end of the file

data LexerWarning

Defines a lexer warning

Constructors

LexerWarning FilePos LexerWarningInfo
The lexer file position and warning information

data LexerWarningInfo

Defines the different lexer warning types

Constructors

TabCharacter
Tab character was encountered

NestedComment FilePos
Nested comment was encountered

UtfChar
The UTF BOM character was found

CommentOperator
Syntax coloring cannot handle names containing --

keepOneTabWarning

Arguments

:: [LexerWarning]
The old warnings

-> [LexerWarning]
The new warnings

Generates a TabCharacter warning

showLexerErrorInfo

Arguments

:: LexerErrorInfo
The error information

-> [String]
The error messages

Converts the error information into a list of error messages

showLexerWarningInfo

Arguments

:: LexerWarningInfo
The warning information

-> [String]
The warning messages

Converts the warning information into a list of warning messages

data Token

The Ampersand token

Constructors

Tok

Fields

tokLex :: Lexeme

The lexeme

tokPos :: FilePos

The position where the lexeme was found

data Lexeme

The Ampersand Lexemes

Constructors

LexSymbol Char
A symbol

LexOperator String
An operator

LexKeyword String
A keyword

LexString String
A quoted string

LexExpl String
An explanation

LexAtom String
An atom

LexDecimal Int
A decimal number

LexOctal Int
An octal number

LexHex Int
A hexadecimal number

LexConId String
An upper case identifier

LexVarId String
A lower case identifier

lexemeText

Arguments

:: Lexeme
The lexeme

-> String
The text contained in the lexeme

Retrieves the text from a lexeme

initPos

Arguments

:: FilePath
The file path

-> FilePos
The initial position

Makes the initial position for a file

data FilePos

Represents a position within a file, including the file path, line and column numbers

Constructors

FilePos FilePath Line Column