Database.Design.Ampersand.Input.ADL1.LexerMessage

Documentation

data LexerError

Defines a lexer error

Constructors

LexerError FilePos LexerErrorInfo
The lexer file position and error information

data LexerErrorInfo

Defines the different lexer error types

Constructors

UnterminatedComment
The comment was not terminated

UnterminatedPurpose
The purpose statement was not terminated

UnterminatedAtom
The atom was not terminated

UnexpectedChar Char
The character was unexpected

NonTerminatedString String
The string was not terminated

TooManyClose Char
In UnexpectedClose, first char is the closing bracket we see, second char is the closing bracket we would like to see first e.g. [(1,3] => UnexpectedClose ']' ... ')'

UnexpectedClose Char FilePos Char
An unexpected closing bracket was found

StillOpenAtEOF [(FilePos, Char)]
Brackets were not closed at the end of the file

data LexerWarning

Defines a lexer warning

Constructors

LexerWarning FilePos LexerWarningInfo
The lexer file position and warning information

data LexerWarningInfo

Defines the different lexer warning types

Constructors

TabCharacter
Tab character was encountered

NestedComment FilePos
Nested comment was encountered

UtfChar
The UTF BOM character was found

CommentOperator
Syntax coloring cannot handle names containing --

keepOneTabWarning

Arguments

:: [LexerWarning]
The old warnings

-> [LexerWarning]
The new warnings

Generates a TabCharacter warning

showLexerErrorInfo

Arguments

:: LexerErrorInfo
The error information

-> [String]
The error messages

Converts the error information into a list of error messages

showLexerWarningInfo

Arguments

:: LexerWarningInfo
The warning information

-> [String]
The warning messages

Converts the warning information into a list of warning messages