Database.Design.Ampersand.Input.ADL1.LexerMonad

Description

Documentation

data LexerMonad a

The type of the Lexer Monad Output monad: [LexerWarning] State monad: FilePos and [Bracket]

getPos :: LexerMonad FilePos

Retrieves the position of a lexer monad

incPos

Arguments

:: Int
The amount of columns read

-> LexerMonad ()
The resulting monad

Increases the position of a lexer monad with a given number of columns

nextPos :: Char -> LexerMonad ()

Increases the position of a lexer monad with a given character

addPos :: Int -> FilePos -> FilePos

Adds one column to the file position

openBracket

Arguments

:: Char
The found bracket

-> LexerMonad ()
The resulting monad

Adds a new opening bracket to the lexer monad

closeBracket

Arguments

:: Char
The found bracket

-> LexerMonad ()
The resulting monad

Adds a new closing bracket to the lexer monad, trying to match an opening one

checkBracketsAtEOF

Arguments

:: LexerMonad ()
The resulting monad

Asserts that the bracket list is empty at the end of the file

lexerError

Arguments

:: LexerErrorInfo
The generated error

-> FilePos
The location where the error is originated

-> LexerMonad a
The resulting monad

Generates a monad with an error message

lexerWarning

Arguments

:: LexerWarningInfo
The generated warning

-> FilePos
The location where the warning is originated

-> LexerMonad ()
The resulting monad

Generates a monad with a warning message

runLexerMonad

Arguments

:: [Options]
The command line options

-> FilePath
The file to be read (used for error messages)

-> LexerMonad a
The lexer monad to run

-> Either LexerError (a, [LexerWarning])
Result is either an error or a result and a list of warnings

Runs the lexer monad

getOpts

Arguments

:: LexerMonad [Options]
The lexer monad with the options

Retrieves the command line options