Database.Design.Ampersand.Input.ADL1.Parser

Documentation

type AmpParser a

Arguments

= ParsecT [Token] FilePos Identity a
The Parsec parser for a list of tokens with a file position.

The Ampersand parser type

pContext

Arguments

:: AmpParser (P_Context, [String])
The result is the parsed context and a list of include filenames

Parses a context

pPopulations

Arguments

:: AmpParser [P_Population]
The population list parser

Parses a list of populations

pTerm

Arguments

:: AmpParser (Term TermPrim)
The term parser

Parses a term

pRule

Arguments

:: AmpParser (Term TermPrim)
The rule parser

Parses a rule