Database.Design.Ampersand.Input.ADL1.ParsingLib

Documentation

type AmpParser a

Arguments

= ParsecT [Token] FilePos Identity a
The Parsec parser for a list of tokens with a file position.

The Ampersand parser type

pIsThere

Arguments

:: AmpParser a
The parser to run

-> AmpParser Bool
The parser with the result

Tries to apply the given parser and returns a parser with a boolean indicating whether it succeeded

optList :: AmpParser [a] -> AmpParser [a]

Optionally applies a list parser, returning an empty list if it doesn't succeed

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

(<|>) :: ParsecT s u m a -> ParsecT s u m a -> ParsecT s u m a

(<$) infixl 4

Arguments

:: a
The value to return

-> AmpParser b
The parser to apply

-> AmpParser a
The result

Applies the given parser and returns the given constructor

(<*>) :: Applicative f => forall a b. f (a -> b) -> f a -> f b

Sequential application.

(<*) :: Applicative f => forall a b. f a -> f b -> f a

Sequence actions, discarding the value of the second argument.

(*>) :: Applicative f => forall a b. f a -> f b -> f b

Sequence actions, discarding the value of the first argument.

(<??>) :: AmpParser a -> AmpParser (a -> a) -> AmpParser a

sepBy :: Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

sepBy1 :: Stream s m t => ParsecT s u m a -> ParsecT s u m sep -> ParsecT s u m [a]

many :: ParsecT s u m a -> ParsecT s u m [a]

many1 :: Stream s m t => ParsecT s u m a -> ParsecT s u m [a]

opt :: AmpParser a -> a -> AmpParser a

try :: ParsecT s u m a -> ParsecT s u m a

choice :: Stream s m t => [ParsecT s u m a] -> ParsecT s u m a

pMaybe

Arguments

:: AmpParser a
The parser to apply

-> AmpParser (Maybe a)
The result

Tries to apply the given parser and encapsulates the result in Maybe

currPos :: AmpParser Origin

posOf :: Show a => AmpParser a -> AmpParser Origin

valPosOf :: Show a => AmpParser a -> AmpParser (a, Origin)

pAtom :: AmpParser String

pConid :: AmpParser String

pString :: AmpParser String

pExpl :: AmpParser String

pVarid :: AmpParser String

pComma :: AmpParser String

pParens :: AmpParser a -> AmpParser a

pBraces :: AmpParser a -> AmpParser a

pBrackets :: AmpParser a -> AmpParser a

pChevrons :: AmpParser a -> AmpParser a

pKey :: String -> AmpParser String

pOperator :: String -> AmpParser String

pDash :: AmpParser String

pSemi :: AmpParser String

pColon :: AmpParser String

pZero :: AmpParser String

pOne :: AmpParser String

pInteger :: AmpParser Int