Database.Design.Ampersand.Input.Parsing

Documentation

parseADL

Arguments

:: Options
The options given through the command line

-> FilePath
The path of the file to be parsed

-> IO (Guarded P_Context)
The resulting context

Parse an Ampersand file and all transitive includes

parseADL1pExpr

Arguments

:: String
The string to be parsed

-> FilePath
The name of the file (used for error messages)

-> Either String (Term TermPrim)
The result: Either an error message, or a good result

Parses an isolated ADL1 expression string

parseRule

Arguments

:: String
The string to be parsed

-> Term TermPrim
The resulting rule

Parses an isolated rule In order to read derivation rules, we use the Ampersand parser. Since it is applied on static code only, error messagea may be produced as fatals.

parseCtx

Arguments

:: FilePath
The file name (used for error messages)

-> String
The string to be parsed

-> Guarded (P_Context, [String])
The context and a list of included files

Parses an Ampersand context

runParser

Arguments

:: AmpParser a
The parser to run

-> FilePath
Name of the file (for error messages)

-> String
String to parse

-> Guarded a
The result

Runs the given parser