Database.Design.Ampersand.Misc.Languages

Description

This module does some string manipulation based on natural languages

Documentation

data Lang

Constructors

Dutch

English

allLangs :: [Lang]

plural :: Lang -> String -> String

Returns the plural of a given word based on a specific language