Database.Design.Ampersand.Misc.Options

Documentation

data Options

This data constructor is able to hold all kind of information that is useful to express what the user would like Ampersand to do.

Constructors

Options

Fields

showVersion :: Bool

preVersion :: String

postVersion :: String

showHelp :: Bool

verboseP :: Bool

development :: Bool

validateSQL :: Bool

validateEdit :: Maybe String

genPrototype :: Bool

dirPrototype :: String

allInterfaces :: Bool

dbName :: String

genAtlas :: Bool

namespace :: String

autoRefresh :: Maybe Int

testRule :: Maybe String

customCssFile :: Maybe FilePath

importfile :: FilePath

fileformat :: Maybe FileFormat

theme :: DocTheme

genXML :: Bool

genFSpec :: Bool

diag :: Bool

fspecFormat :: FSpecFormat

genGraphics :: Bool

genEcaDoc :: Bool

proofs :: Bool

haskell :: Bool

dirOutput :: String

outputfile :: String

crowfoot :: Bool

blackWhite :: Bool

doubleEdges :: Bool

showPredExpr :: Bool

noDiagnosis :: Bool

diagnosisOnly :: Bool

genLegalRefs :: Bool

genUML :: Bool

genFPAChap :: Bool

genFPAExcel :: Bool

genStaticFiles :: Bool

genBericht :: Bool

language :: Maybe Lang

dirExec :: String

ampersandDataDir :: FilePath

progrName :: String

fileName :: FilePath

baseName :: String

logName :: FilePath

genTime :: LocalTime

export2adl :: Bool

test :: Bool

genASTTables :: Bool

genASTFile :: Bool

genGenericsFile :: Bool

genGenericTables :: Bool

metaTablesHaveUnderscore :: Bool

pangoFont :: String

sqlHost :: String

sqlLogin :: String

sqlPwd :: String

oldNormalizer :: Bool

newFrontend :: Bool

getOptions :: IO Options

usageInfo' :: Options -> String

verboseLn :: Options -> String -> IO ()

verbose :: Options -> String -> IO ()

data FSpecFormat

Constructors

FPandoc

Fasciidoc

Fcontext

Fdocbook

Fhtml

FLatex

Fman

Fmarkdown

Fmediawiki

Fopendocument

Forg

Fplain

Frst

Frtf

Ftexinfo

Ftextile

data FileFormat

Constructors

Adl1Format

Adl1PopFormat

data DocTheme

Constructors

DefaultTheme

ProofTheme

StudentTheme

StudentDesignerTheme

DesignerTheme

helpNVersionTexts :: String -> Options -> [String]