Database.Design.Ampersand.Output.ToPandoc.SharedAmongChapters

Documentation

bulletList :: [Blocks] -> Blocks

math :: String -> Inlines

module Data.Monoid

module Database.Design.Ampersand.Basics

module Database.Design.Ampersand.FSpec

module Database.Design.Ampersand.Misc

module Database.Design.Ampersand.Core.AbstractSyntaxTree

module Database.Design.Ampersand.ADL1

module Database.Design.Ampersand.Output.PandocAux

data Chapter

Constructors

Intro

SharedLang

Diagnosis

ConceptualAnalysis

ProcessAnalysis

DataAnalysis

SoftwareMetrics

EcaRules

Interfaces

FunctionPointAnalysis

Glossary

chaptersInDoc :: Options -> [Chapter]

Define the order of the chapters in the document.

chptHeader :: Lang -> Chapter -> Blocks

This function returns a header of a chapter

chptTitle :: Lang -> Chapter -> Inlines

class Xreferencable a where

Minimal complete definition

xLabel

Methods

xLabel :: a -> String

xRefReference :: Options -> a -> Inlines

xrefLabel :: a -> Inlines

showImage :: Options -> Picture -> Inlines

canXRefer :: Options -> Bool

data Purpose

Explanation is the intended constructor. It explains the purpose of the object it references. The enrichment process of the parser must map the names (from PPurpose) to the actual objects

Constructors

Expl

Fields

explPos :: Origin

The position in the Ampersand script of this purpose definition

explObj :: ExplObj

The object that is explained.

explMarkup :: A_Markup

This field contains the text of the explanation including language and markup info.

explUserdefd :: Bool

Is this purpose defined in the script?

explRefIds :: [String]

The references of the explaination

purposes2Blocks :: Options -> [Purpose] -> Blocks

isMissing :: Maybe Purpose -> Bool

lclForLang :: Lang -> TimeLocale

dpRule' :: FSpec -> [Rule] -> Int -> [A_Concept] -> [Declaration] -> ([(Inlines, [Blocks])], Int, [A_Concept], [Declaration])

relsInThemes :: FSpec -> [Declaration]

inlineIntercalate :: Inlines -> [Inlines] -> Inlines

data ThemeContent

This function orders the content to print by theme. It returns a list of tripples by theme. The last tripple might not have a theme, but will contain everything that isn't handled in a specific theme.

Constructors

Thm

Fields

themeNr :: Int

patOfTheme :: Maybe Pattern

rulesOfTheme :: [Numbered RuleCont]

dclsOfTheme :: [Numbered DeclCont]

cptsOfTheme :: [Numbered CptCont]

orderingByTheme :: FSpec -> [ThemeContent]

data Numbered t

Constructors

Nr

Fields

theNr :: Int

theLoad :: t

data RuleCont

Constructors

CRul

Fields

cRul :: Rule

cRulPurps :: [Purpose]

cRulMeaning :: Maybe A_Markup

data DeclCont

Constructors

CDcl

Fields

cDcl :: Declaration

cDclPurps :: [Purpose]

cDclMeaning :: Maybe A_Markup

cDclPairs :: [Paire]

data CptCont

Constructors

CCpt

Fields

cCpt :: A_Concept

cCptDefs :: [ConceptDef]

cCptPurps :: [Purpose]

plainText :: String -> Blocks

newtype NLString

Constructors

NL String

newtype ENString

Constructors

EN String

type LocalizedStr = (NLString, ENString)

localize :: Lang -> LocalizedStr -> String

sortWith :: Ord b => (a -> b) -> [a] -> [a]

The sortWith function sorts a list of elements using the user supplied function to project something out of each element