Database.Design.Ampersand.Basics.Unique

Documentation

class (Typeable e, Eq e) => Unique e where

In the context of the haskell code, things can be Unique.

Minimal complete definition

showUnique

Methods

self :: e -> UniqueObj e

a representation of a unique thing

uniqueShow :: Bool -> e -> String

representation of a Unique thing into a string.

showUnique :: e -> String

A function to show a unique instance. It is the responsability of the instance definition to make sure that for every a, b of an individual type: a == b == showUnique a == showUnique b

class Named a where

anything could have some label, can't it?

Methods

name :: a -> String